Split the ops block of AppModel.swift into five extension files
The contiguous 2,900-line stretch from the model router through the
account routing verbs moves, verbatim, into extension files: ModelRouter,
SendPipeline (with the quit-time drain), Commands (slash execution,
rewind, permission decisions, model switching), TeamOps (manifests, run
lifecycle verbs, dispatch decisions), and RunQueueAndActivity (queue
control, persisted-run restoration, activity actions, worktree git ops,
schedule validation). The private members the moved code shares with the
class widen to internal with the extension-files breadcrumb; the one
stored property in the stretch (pendingProviderSwitch) returns to the
class body, where stored state must live.

Behavior-identical file split; the full suite passes unchanged.
