+++ Locus/AppModel.swift
@@ -67,6 +67,7 @@ final class AppModel: ObservableObject {
     let teamRunLive = TeamRunLiveModel()
     private var teamRunLiveBridge: AnyCancellable?
     @Published var activeTaskRecord: TaskRecord?  // internal(for: AppModel+UITestFixtures)
+    var pendingProviderSwitch: (accountID: UUID?, model: String)?  // internal(for: AppModel extension files)
     let landingFlow = LandingFlowModel()
     private var landingFlowBridge: AnyCancellable?
     let runs = OrchestrationRunsModel()
@@ -125,7 +126,7 @@ final class AppModel: ObservableObject {
     @Published var isBusy = false
     /// A short, truthful description of where an in-flight steering request
     /// is waiting. It is cleared when the direction joins the active turn.
-    @Published private(set) var steeringState: String?
+    @Published var steeringState: String?  // internal(for: AppModel extension files)
     @Published var selectedMode: WorkMode = .work {
         didSet {
             // Changing modes is taking a stance on what happens next, so a
@@ -253,8 +254,8 @@ final class AppModel: ObservableObject {
     @Published private(set) var launchAtLoginError: String?
     @Published private(set) var automaticInspectorPrompt: AutomaticInspectorPrompt?
     @Published var usageDashboardPresented = false
-    @Published private(set) var lastModelRoutingDecision: ModelRoutingDecision?
-    @Published private(set) var modelRouterMessage = "No scorecard has been run yet."
+    @Published var lastModelRoutingDecision: ModelRoutingDecision?  // internal(for: AppModel extension files)
+    @Published var modelRouterMessage = "No scorecard has been run yet."  // internal(for: AppModel extension files)
     @Published private(set) var proxyHealthRecords: [ProxyHealthRecord] = []
     @Published private(set) var proxyHealthMessage = "Proxy health has not been checked yet."
     @Published private(set) var isCheckingProxyHealth = false
@@ -313,16 +314,16 @@ final class AppModel: ObservableObject {
     let sessionOverview = SessionStateEmitter()
 
     let backend: BackendService  // internal(for: AppModel extension files)
-    private let providerCredentialWriter: (String, String) -> Bool
+    let providerCredentialWriter: (String, String) -> Bool  // internal(for: AppModel extension files)
     private let backendProcess = BackendProcess()
     var taskWorkers: [String: ChatWorkerRuntime] = [:]  // internal(for: AppModel extension files)
-    private var chatAdmissionQueue = ChatAdmissionQueue()
-    private var pendingChatTurns: [String: Task<Void, Never>] = [:]
-    private var pendingChatTurnTokens: [String: UUID] = [:]
+    var chatAdmissionQueue = ChatAdmissionQueue()  // internal(for: AppModel extension files)
+    var pendingChatTurns: [String: Task<Void, Never>] = [:]  // internal(for: AppModel extension files)
+    var pendingChatTurnTokens: [String: UUID] = [:]  // internal(for: AppModel extension files)
     private var pendingSimulatorActions: [String: (sessionID: String, task: Task<Void, Never>)] = [:]
     /// Backends that refused the browser handshake because a turn was running.
     private var pendingBrowserCapabilityTransports: [BackendService] = []
-    private var conversationBackend: BackendService {
+    var conversationBackend: BackendService {  // internal(for: AppModel extension files)
         taskWorkers[currentSessionID]?.service ?? backend
     }
 
@@ -330,7 +331,7 @@ final class AppModel: ObservableObject {
     /// back to the worker that owns the run; sending them to the main control
     /// service can update the shared run database without interrupting the
     /// model call, which makes a cancelled approval reappear on reconnect.
-    private func orchestrationBackend(for runID: String) -> BackendService {
+    func orchestrationBackend(for runID: String) -> BackendService {
         guard let sessionID = Self.orchestrationOwnerSessionID(
             for: runID,
             currentSessionID: currentSessionID,
@@ -419,25 +420,25 @@ final class AppModel: ObservableObject {
     private var runtimeRecoveryAttempt = 0
     private var proxyHealthMonitorTask: Task<Void, Never>?
     private var proxyRouteRestartPending = false
-    private var restoredTranscriptContext: String?
+    var restoredTranscriptContext: String?  // internal(for: AppModel extension files)
     private var pendingDeletedChat: DeletedChatUndo?
     private var profilePersistenceTask: Task<Void, Never>?
     private var settingsPersistenceTask: Task<Void, Never>?
-    private var promptHistoryCursor: Int?
-    private var stashedDraft: String?
+    var promptHistoryCursor: Int?  // internal(for: AppModel extension files)
+    var stashedDraft: String?  // internal(for: AppModel extension files)
     private var pendingSessionReset = false
     /// Whether the turn in flight rewrote the todo list. The approval prompt
     /// is offered only for turns that actually produced a plan — a Plan-mode
     /// chat answer must not re-offer a plan left over from an earlier run.
-    private var planTodosChangedThisTurn = false
-    private var planReadyThisTurn = false
+    var planTodosChangedThisTurn = false  // internal(for: AppModel extension files)
+    var planReadyThisTurn = false  // internal(for: AppModel extension files)
     /// What the user actually dispatched, kept separate from the live picker
     /// so a mid-run mode change cannot relabel the completion marker or alter
     /// plan reconciliation.
     var turnDispatchedMode: WorkMode?
-    private var turnDispatchedTeamRunID: String?
+    var turnDispatchedTeamRunID: String?  // internal(for: AppModel extension files)
     /// Client-side fallback for agents from before `turn_done.duration_ms`.
-    private var turnStartedAt: Date?
+    var turnStartedAt: Date?  // internal(for: AppModel extension files)
     /// The window during which file activity counts as this run's output, and
     /// the session it belongs to.
     ///
@@ -454,10 +455,10 @@ final class AppModel: ObservableObject {
     /// Present only for solo turns that the optional model router prepared.
     /// Keeping these per session matters because a routed chat can finish after
     /// the user has switched to, or started, another conversation.
-    private var automaticModelRoutingTurns: [String: ModelRoutingPreparedTurn] = [:]
+    var automaticModelRoutingTurns: [String: ModelRoutingPreparedTurn] = [:]  // internal(for: AppModel extension files)
     /// Keeps workspace persistence pinned to the user's manual choice while a
     /// slower hosted-provider switch restores that choice after opt-out.
-    private var isRestoringManualModelRoute = false
+    var isRestoringManualModelRoute = false  // internal(for: AppModel extension files)
     /// Turning Just Chat off returns to the last mode that could act on the
     /// workspace instead of always making the user reselect Build or Plan.
     private var lastAgenticMode: WorkMode = .work
@@ -470,13 +471,13 @@ final class AppModel: ObservableObject {
     /// into an "implement this plan?" offer. Internal so tests can dispatch
     /// turns without a live backend.
     var turnDispatchedInPlanMode = false
-    private var pendingRetry = false
+    var pendingRetry = false  // internal(for: AppModel extension files)
     /// Stop & Send is deliberately not part of the ordinary queue: it must
     /// wait for the interrupted turn's terminal event before it can create a
     /// fresh provider turn and conversation-history boundary.
-    private var pendingStopAndSend: String?
+    var pendingStopAndSend: String?  // internal(for: AppModel extension files)
     private var pendingCheckpointRestore: SessionCheckpoint?
-    private var pendingRewindDraft: String?
+    var pendingRewindDraft: String?  // internal(for: AppModel extension files)
     private var pendingWorkspacePath: String?
     private var workspaceToOpenAfterReconnect: String?
     private var appliedWorkspacePath: String?
@@ -493,7 +494,7 @@ final class AppModel: ObservableObject {
     /// credential file: a test must not read — or delete — the secrets of
     /// whoever is running the suite.
     let persistenceEnabled: Bool
-    private let isUITesting: Bool
+    let isUITesting: Bool  // internal(for: AppModel extension files)
     private var isShuttingDown = false
     private var settingsUpdatePreparation: (
         id: UUID,
@@ -2267,2917 +2268,6 @@ final class AppModel: ObservableObject {
         gitWorkspace.refreshBranch()
     }
 
-    // MARK: - Optional solo model router
-
-    func setAutomaticModelRoutingEnabled(_ enabled: Bool) {
-        guard enabled != settings.automaticModelRoutingEnabled else { return }
-        if enabled {
-            rememberManualModelRoute(accountID: activeAccount?.id, model: selectedModel)
-            settings.automaticModelRoutingEnabled = true
-            modelRouterMessage = "Ready — the next solo message will use the scorecard."
-            refreshModelRouterScorecard()
-        } else {
-            isRestoringManualModelRoute = true
-            settings.automaticModelRoutingEnabled = false
-            modelRouterMessage = "Automatic routing is off. Restoring the manual route…"
-            Task { [weak self] in
-                guard let self else { return }
-                await restoreManualModelRoute()
-                isRestoringManualModelRoute = false
-            }
-        }
-    }
-
-    func setAutomaticModelRoutingAllowHosted(_ allowed: Bool) {
-        settings.automaticModelRoutingAllowHosted = allowed
-        modelRouterMessage = allowed
-            ? "Hosted accounts are eligible for future solo messages."
-            : "Only models on this Mac are eligible."
-        refreshModelRouterScorecard()
-    }
-
-    func setModelRoutingPolicy(_ policy: ModelRoutingPolicy) {
-        settings.modelRoutingPolicyRaw = policy.rawValue
-        refreshModelRouterScorecard()
-    }
-
-    /// Re-scores the available routes for a generic task without changing the
-    /// active model. The Router inspector uses this after policy edits and on
-    /// explicit refresh.
-    func refreshModelRouterScorecard() {
-        let candidates = automaticModelRouteCandidates(requiresVision: false)
-        guard !candidates.isEmpty else {
-            lastModelRoutingDecision = nil
-            modelRouterMessage = settings.automaticModelRoutingAllowHosted
-                ? "No ready model routes are available."
-                : "No visible local models are available. Hosted routing is off."
-            return
-        }
-        modelRouterMessage = "Scoring \(candidates.count) eligible route\(candidates.count == 1 ? "" : "s")…"
-        Task { [weak self] in
-            guard let self else { return }
-            do {
-                let decision = try await requestModelRoutingDecision(
-                    candidates: candidates,
-                    tags: ["general"]
-                )
-                lastModelRoutingDecision = decision
-                modelRouterMessage = decision.reason
-            } catch {
-                modelRouterMessage = "Could not load scorecards: \(error.localizedDescription)"
-            }
-        }
-    }
-
-    static func modelRoutingTags(for text: String, mode: WorkMode) -> [String] {
-        let lower = text.lowercased()
-        var tags: [String] = []
-        func add(_ tag: String, when condition: Bool) {
-            if condition, !tags.contains(tag) { tags.append(tag) }
-        }
-        add("coding", when: mode != .ask || [
-            "code", "function", "class", "api", "compile", "refactor", ".swift",
-            ".py", ".js", ".ts", ".rs", ".go",
-        ].contains(where: lower.contains))
-        add("debugging", when: ["bug", "debug", "crash", "error", "failing", "fix"]
-            .contains(where: lower.contains))
-        add("testing", when: ["test", "spec", "verify", "regression"]
-            .contains(where: lower.contains))
-        add("review", when: ["review", "audit", "security", "risk"]
-            .contains(where: lower.contains))
-        add("research", when: ["research", "compare", "sources", "browse", "latest"]
-            .contains(where: lower.contains))
-        add("writing", when: ["write", "rewrite", "draft", "summarize", "explain"]
-            .contains(where: lower.contains))
-        add("long_context", when: text.count > 12_000)
-        if tags.isEmpty { tags.append("general") }
-        return Array(tags.prefix(24))
-    }
-
-    private func prepareAutomaticModelRoute(
-        text: String,
-        mode: WorkMode,
-        requiresVision: Bool
-    ) async -> ModelRoutingPreparedTurn? {
-        let candidates = automaticModelRouteCandidates(requiresVision: requiresVision)
-        guard !candidates.isEmpty else {
-            modelRouterMessage = "No eligible route; using the manual model."
-            return nil
-        }
-        let tags = Self.modelRoutingTags(for: text, mode: mode)
-        do {
-            let decision = try await requestModelRoutingDecision(
-                candidates: candidates,
-                tags: tags
-            )
-            lastModelRoutingDecision = decision
-            modelRouterMessage = decision.reason
-            guard let route = candidates.first(where: { $0.id == decision.selectedID }) else {
-                return nil
-            }
-            guard await applyAutomaticModelRoute(route) else {
-                modelRouterMessage = "The selected route was unavailable; using the manual model."
-                return nil
-            }
-            let score = decision.candidates.first(where: { $0.routeID == route.id })?.score
-            showToast(
-                "Auto route: \(route.name)"
-                    + (score.map { " · \(Int($0.rounded()))" } ?? "")
-            )
-            return ModelRoutingPreparedTurn(routeID: route.id, tags: tags, local: route.local)
-        } catch {
-            modelRouterMessage = "Router unavailable; using the manual model: \(error.localizedDescription)"
-            return nil
-        }
-    }
-
-    private func requestModelRoutingDecision(
-        candidates: [AutomaticModelRouteCandidate],
-        tags: [String]
-    ) async throws -> ModelRoutingDecision {
-        try await backend.post(
-            "/api/model-router/decision",
-            body: [
-                "tags": tags,
-                "weights": settings.resolvedModelRoutingPolicy.weights,
-                "candidates": candidates.map(\.payload),
-            ],
-            as: ModelRoutingDecision.self
-        )
-    }
-
-    private func automaticModelRouteCandidates(
-        requiresVision: Bool
-    ) -> [AutomaticModelRouteCandidate] {
-        var routes: [AutomaticModelRouteCandidate] = localModels.compactMap { model in
-            if requiresVision, model.visionCapable == false { return nil }
-            let id = Self.modelRouteID(accountID: nil, model: model.name)
-            let aliases = agentProfiles.compactMap { profile -> String? in
-                guard profile.route.accountID == nil,
-                      profile.model.caseInsensitiveCompare(model.name) == .orderedSame
-                else { return nil }
-                return profile.id.uuidString
-            }
-            return AutomaticModelRouteCandidate(
-                id: id,
-                name: "\(model.name) · Local",
-                model: model.name,
-                provider: "ollama",
-                accountID: nil,
-                local: true,
-                metering: "self_hosted",
-                memoryBytes: model.size,
-                current: activeAccount == nil
-                    && selectedModel.caseInsensitiveCompare(model.name) == .orderedSame,
-                sampleIDs: [id] + aliases
-            )
-        }
-        guard settings.automaticModelRoutingAllowHosted else { return routes }
-        for account in providerAccounts where account.hasKey {
-            guard accountStatus[account.id]?.isHealthy == true else { continue }
-            let model = account.preferredModel.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !model.isEmpty else { continue }
-            let id = Self.modelRouteID(accountID: account.id, model: model)
-            let aliases = agentProfiles.compactMap { profile -> String? in
-                guard profile.route.accountID == account.id,
-                      profile.model.caseInsensitiveCompare(model) == .orderedSame
-                else { return nil }
-                return profile.id.uuidString
-            }
-            let subscription = account.kind == .chatGPT || account.kind == .kimiCode
-            routes.append(AutomaticModelRouteCandidate(
-                id: id,
-                name: "\(model) · \(account.shortName)",
-                model: model,
-                provider: account.kind.rawValue,
-                accountID: account.id,
-                local: false,
-                metering: subscription ? "subscription" : "metered",
-                memoryBytes: 0,
-                current: activeAccount?.id == account.id
-                    && selectedModel.caseInsensitiveCompare(model) == .orderedSame,
-                sampleIDs: [id] + aliases
-            ))
-        }
-        return routes
-    }
-
-    private static func modelRouteID(accountID: UUID?, model: String) -> String {
-        let source = accountID?.uuidString.lowercased() ?? "ollama"
-        return "model-route:\(source):\(model.lowercased())"
-    }
-
-    private func applyAutomaticModelRoute(_ route: AutomaticModelRouteCandidate) async -> Bool {
-        let previousAccountID = settings.activeAccountID
-        let previousProvider = settings.provider
-        let previousModel = selectedModel
-        let sourceChanged = route.accountID?.uuidString != settings.activeAccountID
-        if let accountID = route.accountID {
-            guard providerAccounts.contains(where: { $0.id == accountID }) else { return false }
-            settings.activeAccountID = accountID.uuidString
-            settings.provider = .remote
-            if !sourceChanged { return true }
-            guard await applyProvider(announce: false) else {
-                settings.activeAccountID = previousAccountID
-                settings.provider = previousProvider
-                _ = await applyProvider(announce: false)
-                return false
-            }
-            return true
-        }
-
-        settings.activeAccountID = nil
-        settings.provider = .ollama
-        if sourceChanged, !(await applyProvider(announce: false)) {
-            settings.activeAccountID = previousAccountID
-            settings.provider = previousProvider
-            _ = await applyProvider(announce: false)
-            return false
-        }
-        do {
-            let state: ConfigStateResponse = try await backend.post(
-                "/api/config", body: ["model": route.model], as: ConfigStateResponse.self
-            )
-            if let info = state.sessionInfo { sessionInfo = info }
-            let transport = conversationBackend
-            if transport !== backend {
-                let _: ConfigStateResponse = try await transport.post(
-                    "/api/config", body: ["model": route.model], as: ConfigStateResponse.self
-                )
-            }
-            return true
-        } catch {
-            settings.activeAccountID = previousAccountID
-            settings.provider = previousProvider
-            _ = await applyProvider(announce: false)
-            if previousAccountID == nil, !previousModel.isEmpty {
-                let _: ConfigStateResponse? = try? await backend.post(
-                    "/api/config", body: ["model": previousModel], as: ConfigStateResponse.self
-                )
-            }
-            return false
-        }
-    }
-
-    private func rememberManualModelRoute(accountID: UUID?, model: String) {
-        let trimmed = model.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty, trimmed != "No model" else { return }
-        settings.modelRouterFallbackAccountID = accountID?.uuidString
-        settings.modelRouterFallbackModel = trimmed
-    }
-
-    private func restoreManualModelRoute() async {
-        let model = settings.modelRouterFallbackModel
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !model.isEmpty else {
-            modelRouterMessage = "Automatic routing is off."
-            return
-        }
-        let accountID = settings.modelRouterFallbackAccountID.flatMap(UUID.init(uuidString:))
-        let candidate = AutomaticModelRouteCandidate(
-            id: Self.modelRouteID(accountID: accountID, model: model),
-            name: model,
-            model: model,
-            provider: accountID == nil ? "ollama" : "account",
-            accountID: accountID,
-            local: accountID == nil,
-            metering: accountID == nil ? "self_hosted" : "metered",
-            memoryBytes: 0,
-            current: false,
-            sampleIDs: []
-        )
-        modelRouterMessage = await applyAutomaticModelRoute(candidate)
-            ? "Automatic routing is off. Manual route restored."
-            : "Automatic routing is off, but the saved manual route is unavailable."
-    }
-
-    private func discardAutomaticModelRoutingTurn(
-        for sessionID: String,
-        matching prepared: ModelRoutingPreparedTurn?
-    ) {
-        guard let prepared,
-              automaticModelRoutingTurns[sessionID]?.id == prepared.id
-        else { return }
-        automaticModelRoutingTurns.removeValue(forKey: sessionID)
-    }
-
-    private func recordAutomaticModelRoutingOutcome(
-        sessionID: String,
-        reason: String,
-        backendDurationMilliseconds: Int?
-    ) {
-        guard let routed = automaticModelRoutingTurns.removeValue(forKey: sessionID) else {
-            return
-        }
-        let startedAt = taskWorkers[sessionID]?.startedAt
-            ?? (currentSessionID == sessionID ? turnStartedAt : nil)
-        let elapsed = startedAt.map { Int(Date().timeIntervalSince($0) * 1_000) } ?? 0
-        let duration = max(backendDurationMilliseconds ?? elapsed, 0)
-        Task { [weak self] in
-            guard let self else { return }
-            let _: SimpleActionResponse? = try? await backend.post(
-                "/api/model-router/sample",
-                body: [
-                    "route_id": routed.routeID,
-                    "tags": routed.tags,
-                    "reliable": reason == "complete",
-                    "latency_ms": duration,
-                    "estimated_cost": 0,
-                    "local": routed.local,
-                    "evaluation": false,
-                ],
-                as: SimpleActionResponse.self
-            )
-        }
-    }
-
-    func send(_ rawText: String) {
-        send(rawText, preservingDraftOnFailure: true)
-    }
-
-    private func send(
-        _ rawText: String,
-        preservingDraftOnFailure: Bool,
-        requeueingOnFailure: Bool = false,
-        includeAttachments: Bool = true,
-        automaticRoutingPrepared: Bool = false,
-        preparedModelRoute: ModelRoutingPreparedTurn? = nil
-    ) {
-        let text = rawText.trimmingCharacters(in: .whitespacesAndNewlines)
-        let availableAttachments = includeAttachments ? availableChatAttachments : []
-        let hasChatAttachments = !availableAttachments.isEmpty
-        guard !text.isEmpty || hasChatAttachments else { return }
-
-        // Slash commands that Locus can run itself execute immediately, even
-        // mid-run; anything else starting with "/" goes to the agent verbatim.
-        if !text.isEmpty, let command = SlashCommand.command(invokedBy: text) {
-            if draftText.trimmingCharacters(in: .whitespacesAndNewlines) == text {
-                draftText = ""
-            }
-            recordPrompt(text)
-            execute(command, argument: SlashCommand.argument(in: text))
-            return
-        }
-
-        if isBusy || hasPendingPermission {
-            if hasChatAttachments {
-                showToast("Wait for the current reply before sending attachments")
-                return
-            }
-            queuedMessages.append(text)
-            taskWorkers[currentSessionID]?.queuedMessages = queuedMessages
-            if draftText.trimmingCharacters(in: .whitespacesAndNewlines) == text {
-                draftText = ""
-            }
-            showToast(
-                hasPendingPermission
-                    ? "Queued — answer the permission request to continue"
-                    : "Queued — sends when this turn finishes"
-            )
-            return
-        }
-        guard isAgentOnline else {
-            stashUnsent(text, requeue: requeueingOnFailure, preserveDraft: preservingDraftOnFailure)
-            return
-        }
-
-        let isSlashPassthrough = SlashCommand.query(from: text) != nil
-        // Capture the mode before any asynchronous context work. A user can
-        // change the picker while that work is pending; the dispatched turn
-        // must keep the safety contract it started with.
-        let dispatchedMode = selectedMode
-        let teamMention = TeamMentionResolver.selection(
-            in: text,
-            profiles: agentProfiles,
-            teams: agentTeams
-        )
-        let wantsTeam = dispatchedMode != .ask
-            && !isSlashPassthrough
-            && (selectedAgentTeamID != nil || teamMention.agent != nil || teamMention.team != nil)
-        let dispatchedTeam = wantsTeam ? teamManifest(for: text) : nil
-        if wantsTeam, dispatchedTeam == nil { return }
-        let dispatchedSoloSwarm = dispatchedTeam == nil
-            && selectedAgentTeamID == nil
-            && dispatchedMode != .ask
-            && !isSlashPassthrough
-        if settings.automaticModelRoutingEnabled,
-           !automaticRoutingPrepared,
-           !isSlashPassthrough,
-           dispatchedTeam == nil
-        {
-            isBusy = true
-            modelRouterMessage = "Choosing a model for this message…"
-            let requiresVision = availableAttachments.contains {
-                $0.kind == .image || $0.kind == .applicationSnapshot
-            }
-            Task { [weak self] in
-                guard let self else { return }
-                let route = await prepareAutomaticModelRoute(
-                    text: text,
-                    mode: dispatchedMode,
-                    requiresVision: requiresVision
-                )
-                isBusy = false
-                send(
-                    rawText,
-                    preservingDraftOnFailure: preservingDraftOnFailure,
-                    requeueingOnFailure: requeueingOnFailure,
-                    includeAttachments: includeAttachments,
-                    automaticRoutingPrepared: true,
-                    preparedModelRoute: route
-                )
-            }
-            return
-        }
-        // Agent-side slash commands never receive attachments (the server
-        // routes them past the turn machinery), so dispatching any would
-        // silently drop them — keep the chips for the next real message.
-        let dispatchedAttachments = isSlashPassthrough && dispatchedMode != .ask
-            ? [] : availableAttachments
-        let messageText = text.isEmpty ? "Please analyze the attached files." : text
-        let dispatchedSessionID = currentSessionID
-        let dispatchedWorkspaceRoot = workspacePath
-        let dispatchedExecutionPath = activeTaskRecord?.executionPath ?? dispatchedWorkspaceRoot
-        let dispatchedEnvironment = currentExecutionEnvironment
-        let dispatchedContextFiles = contextFiles
-        let dispatchedLiveApplication = dispatchedMode == .ask ? nil
-            : liveApplicationTargets[dispatchedSessionID].flatMap {
-                applicationContext.isConnected($0) ? $0 : nil
-            }
-        let dispatchedSimulator = dispatchedMode == .ask
-            ? nil : simulatorControl.target(for: dispatchedSessionID)
-        let dispatchedRestoredContext = isSlashPassthrough ? nil : restoredTranscriptContext
-        if !isSlashPassthrough { restoredTranscriptContext = nil }
-
-        isBusy = true
-        turnStartedAt = Date()
-        automaticModelRoutingTurns[dispatchedSessionID] = preparedModelRoute
-        planApprovalPending = false
-        planTodosChangedThisTurn = false
-        planReadyThisTurn = false
-        clearPendingQuestion()
-        // Agent-side slash commands (/init and friends) may write todos, but
-        // running one is housekeeping, never a plan worth offering to build.
-        turnDispatchedInPlanMode = dispatchedMode == .plan && !isSlashPassthrough
-        turnDispatchedMode = isSlashPassthrough ? nil : dispatchedMode
-        turnDispatchedTeamRunID = dispatchedTeam?["run_id"] as? String
-        let oneMessageSnapshotIDs = Self.attachmentIDsToClear(
-            dispatchedAttachments,
-            deliverySucceeded: true
-        ).subtracting(
-            Self.attachmentIDsToClear(dispatchedAttachments, deliverySucceeded: false)
-        )
-        if !dispatchedAttachments.isEmpty {
-            let sentIDs = Self.attachmentIDsToClear(
-                dispatchedAttachments,
-                deliverySucceeded: false
-            )
-            chatAttachments.removeAll { sentIDs.contains($0.id) }
-            chatAttachmentNotice = nil
-        }
-        let attachmentLine = dispatchedAttachments.isEmpty
-            ? nil
-            : "Attached: \(dispatchedAttachments.map(\.name).joined(separator: ", "))"
-        let visibleText = [text.nilIfEmpty, attachmentLine]
-            .compactMap { $0 }
-            .joined(separator: "\n\n")
-        let teamRunID = (dispatchedTeam?["run_id"] as? String)?.nilIfEmpty
-        let reservedRunID = teamRunID ?? UUID().uuidString
-        let visibleBlock = ChatBlock(kind: .user, text: visibleText, runID: reservedRunID)
-        blocks.append(visibleBlock)
-        if let info = sessionInfo, sessionOverview.activeSessionID != info.sessionID {
-            activateSessionOverview(info)
-        }
-        sessionOverview.emit(.message(role: .user, at: Self.sessionTimestamp))
-        sessionOverview.emit(.status(
-            status: .running,
-            reason: nil,
-            at: Self.sessionTimestamp
-        ))
-        beginSessionFileCapture()
-        // Agent-side slash commands go out as raw text: no context pack, so
-        // none of it counts as provided.
-        let providedItems = Self.providedSourceItems(
-            attachments: dispatchedAttachments,
-            contextFiles: isSlashPassthrough ? [] : dispatchedContextFiles,
-            mode: dispatchedMode,
-            liveApplication: dispatchedLiveApplication,
-            simulator: dispatchedSimulator
-        )
-        if !providedItems.isEmpty {
-            sessionOverview.emit(.sourceProvided(items: providedItems, at: Self.sessionTimestamp))
-        }
-        if !text.isEmpty { recordPrompt(text) }
-        if draftText.trimmingCharacters(in: .whitespacesAndNewlines) == text {
-            draftText = ""
-        }
-        // Adaptive workers stay inside the ordinary Solo experience. The Runs
-        // inspector still opens automatically for explicit teams only.
-        let opensRuns = dispatchedTeam != nil
-        presentInspectorForSentRequest(
-            isTeam: opensRuns,
-            runID: opensRuns ? reservedRunID : nil
-        )
-        let previousRuntimeState = taskConversationStates[dispatchedSessionID]
-        taskConversationStates[dispatchedSessionID] = TaskConversationState(
-            sessionID: dispatchedSessionID,
-            taskID: activeTaskRecord?.id ?? previousRuntimeState?.taskID,
-            teamID: previousRuntimeState?.teamID,
-            workerID: previousRuntimeState?.workerID,
-            runID: reservedRunID,
-            state: .queued,
-            updatedAt: Date()
-        )
-
-        let pendingTurnToken = UUID()
-        let pendingTurn = Task { [weak self] in
-            guard let self else { return }
-            defer {
-                if self.pendingChatTurnTokens[dispatchedSessionID] == pendingTurnToken {
-                    self.pendingChatTurns.removeValue(forKey: dispatchedSessionID)
-                    self.pendingChatTurnTokens.removeValue(forKey: dispatchedSessionID)
-                }
-            }
-            guard !Task.isCancelled else {
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                return
-            }
-            do {
-                let queuedTeam = dispatchedTeam?["team"] as? [String: Any]
-                let _: OrchestrationRun = try await self.backend.post(
-                    "/api/runs/queue",
-                    body: [
-                        "run_id": reservedRunID,
-                        "session_id": dispatchedSessionID,
-                        "message_id": visibleBlock.id.uuidString,
-                        "workspace_root": dispatchedWorkspaceRoot,
-                        "execution_path": dispatchedExecutionPath,
-                        "request": messageText,
-                        "run_kind": dispatchedTeam == nil ? "solo" : "team",
-                        "team_id": queuedTeam?["id"] as? String ?? "",
-                        "team_name": queuedTeam?["name"] as? String ?? "",
-                        "execution_environment": dispatchedEnvironment.rawValue,
-                        "solo_swarm": dispatchedSoloSwarm,
-                    ],
-                    as: OrchestrationRun.self
-                )
-            } catch {
-                if let previousRuntimeState {
-                    self.taskConversationStates[dispatchedSessionID] = previousRuntimeState
-                } else {
-                    self.taskConversationStates.removeValue(forKey: dispatchedSessionID)
-                }
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                if self.currentSessionID == dispatchedSessionID {
-                    self.isBusy = false
-                    self.turnStartedAt = nil
-                    self.turnDispatchedMode = nil
-                    self.turnDispatchedTeamRunID = nil
-                    self.turnDispatchedInPlanMode = false
-                    self.stashUnsent(
-                        text,
-                        requeue: requeueingOnFailure,
-                        preserveDraft: preservingDraftOnFailure
-                    )
-                } else if requeueingOnFailure,
-                          let runtime = self.taskWorkers[dispatchedSessionID] {
-                    runtime.queuedMessages.insert(text, at: 0)
-                }
-                self.showToast("Could not queue this chat: \(error.localizedDescription)")
-                return
-            }
-            let refreshedContextFiles = dispatchedMode == .ask || dispatchedContextFiles.isEmpty
-                ? dispatchedContextFiles
-                : await Task.detached(priority: .utility) {
-                    dispatchedContextFiles.map(ContextPackLoader.reloadContextReference)
-                }.value
-            guard !Task.isCancelled else {
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                return
-            }
-            let payload = isSlashPassthrough
-                ? text
-                : Self.decoratedPrompt(
-                    messageText,
-                    mode: dispatchedMode,
-                    chatAttachments: dispatchedAttachments,
-                    contextFiles: refreshedContextFiles,
-                    restoredTranscriptContext: dispatchedRestoredContext,
-                    liveApplication: dispatchedLiveApplication,
-                    simulator: dispatchedSimulator
-                )
-            var request: [String: Any] = [
-                "type": "user_message",
-                "text": payload,
-                "mode": dispatchedMode.rawValue,
-            ]
-            if let agentConfig = encodedJSONObject(self.primaryAgentBehavior) {
-                request["agent_config"] = agentConfig
-            }
-            if let dispatchedTeam { request["team"] = dispatchedTeam }
-            if dispatchedTeam == nil { request["run_id"] = reservedRunID }
-            if dispatchedSoloSwarm {
-                request["solo_swarm"] = ["enabled": true]
-            }
-            let imageAttachments: [[String: Any]] = dispatchedAttachments.compactMap {
-                attachment in
-                guard attachment.kind == .image || attachment.kind == .applicationSnapshot,
-                      let data = attachment.imageData,
-                      let mimeType = attachment.mimeType
-                else { return nil }
-                return [
-                    "name": attachment.name,
-                    "mime_type": mimeType,
-                    "data": data.base64EncodedString(),
-                ]
-            }
-            if !imageAttachments.isEmpty { request["attachments"] = imageAttachments }
-            guard let worker = await self.ensureChatWorker(
-                for: dispatchedSessionID,
-                workspaceRoot: dispatchedWorkspaceRoot
-            ) else {
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                if Task.isCancelled { return }
-                if self.currentSessionID == dispatchedSessionID {
-                    self.isBusy = false
-                    self.turnStartedAt = nil
-                    self.turnDispatchedMode = nil
-                    self.turnDispatchedTeamRunID = nil
-                    self.turnDispatchedInPlanMode = false
-                    self.stashUnsent(
-                        text,
-                        requeue: requeueingOnFailure,
-                        preserveDraft: preservingDraftOnFailure
-                    )
-                }
-                return
-            }
-            guard !Task.isCancelled else {
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                self.finishChatRuntime(worker, state: .cancelled)
-                return
-            }
-            worker.dispatchedMode = isSlashPassthrough ? nil : dispatchedMode
-            worker.dispatchedTeamRunID = teamRunID
-            worker.reservedRunID = reservedRunID
-            worker.dispatchedInPlanMode = dispatchedMode == .plan && !isSlashPassthrough
-            guard await self.waitForChatExecutionSlot(worker) else {
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                return
-            }
-            do {
-                let _: OrchestrationRun = try await self.backend.patch(
-                    "/api/runs/\(reservedRunID)/queue",
-                    body: ["action": "admit"],
-                    as: OrchestrationRun.self
-                )
-            } catch {
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                self.finishChatRuntime(worker, state: .failed, error: "The queued run could not start")
-                return
-            }
-            guard worker.service.send(request) else {
-                self.discardAutomaticModelRoutingTurn(
-                    for: dispatchedSessionID,
-                    matching: preparedModelRoute
-                )
-                self.finishChatRuntime(worker, state: .failed, error: "The turn could not be delivered")
-                if self.currentSessionID == dispatchedSessionID {
-                    self.isBusy = false
-                    self.turnStartedAt = nil
-                    self.turnDispatchedMode = nil
-                    self.turnDispatchedTeamRunID = nil
-                    self.turnDispatchedInPlanMode = false
-                    self.stashUnsent(
-                        text,
-                        requeue: requeueingOnFailure,
-                        preserveDraft: preservingDraftOnFailure
-                    )
-                }
-                return
-            }
-            // Appshots are explicit one-message captures. Retain them through
-            // queue and transport failures; clear only after accepted delivery.
-            if self.currentSessionID == dispatchedSessionID, !oneMessageSnapshotIDs.isEmpty {
-                self.chatAttachments.removeAll { oneMessageSnapshotIDs.contains($0.id) }
-                if self.chatAttachments.isEmpty { self.chatAttachmentNotice = nil }
-            }
-            worker.executionState = dispatchedTeam == nil ? .running : .dispatching
-            worker.startedAt = Date()
-            self.updateBackgroundChatState(worker)
-        }
-        pendingChatTurnTokens[dispatchedSessionID] = pendingTurnToken
-        pendingChatTurns[dispatchedSessionID] = pendingTurn
-    }
-
-    private func waitForChatExecutionSlot(_ runtime: ChatWorkerRuntime) async -> Bool {
-        runtime.executionState = .queued
-        runtime.startedAt = nil
-        updateBackgroundChatState(runtime)
-        chatAdmissionQueue.enqueue(runtime.sessionID)
-        while runtime.process.isRunning {
-            if Task.isCancelled {
-                chatAdmissionQueue.remove(runtime.sessionID)
-                return false
-            }
-            let occupied = taskWorkers.values.filter {
-                $0 !== runtime && $0.occupiesExecutionSlot
-            }.count
-            if chatAdmissionQueue.isFirst(runtime.sessionID),
-               occupied < AppSettings.clampMaximumActiveChats(settings.maximumActiveChats),
-               !hasLocalWriterCollision(for: runtime) {
-                chatAdmissionQueue.remove(runtime.sessionID)
-                runtime.executionState = .running
-                runtime.startedAt = Date()
-                updateBackgroundChatState(runtime)
-                return true
-            }
-            try? await Task.sleep(for: .milliseconds(200))
-            if Task.isCancelled {
-                chatAdmissionQueue.remove(runtime.sessionID)
-                return false
-            }
-        }
-        chatAdmissionQueue.remove(runtime.sessionID)
-        return false
-    }
-
-    private func hasLocalWriterCollision(for runtime: ChatWorkerRuntime) -> Bool {
-        guard runtime.dispatchedMode == .work || runtime.dispatchedMode == .grill,
-              runtime.sessionInfo?.environment?["type"] != ChatExecutionEnvironment.worktree.rawValue,
-              let root = runtime.sessionInfo?.environment?["canonical_repository"]
-                ?? runtime.sessionInfo?.workspaceRoot ?? runtime.sessionInfo?.cwd
-        else { return false }
-        let canonical = URL(fileURLWithPath: root).standardizedFileURL.path
-        return taskWorkers.values.contains { other in
-            guard other !== runtime, other.occupiesExecutionSlot,
-                  other.dispatchedMode == .work || other.dispatchedMode == .grill,
-                  other.sessionInfo?.environment?["type"]
-                    != ChatExecutionEnvironment.worktree.rawValue,
-                  let otherRoot = other.sessionInfo?.environment?["canonical_repository"]
-                    ?? other.sessionInfo?.workspaceRoot ?? other.sessionInfo?.cwd
-            else { return false }
-            return URL(fileURLWithPath: otherRoot).standardizedFileURL.path == canonical
-        }
-    }
-
-    func updateBackgroundChatState(_ runtime: ChatWorkerRuntime) {  // internal(for: AppModel extension files)
-        let previous = taskConversationStates[runtime.sessionID]
-        taskConversationStates[runtime.sessionID] = TaskConversationState(
-            sessionID: runtime.sessionID,
-            taskID: runtime.sessionInfo?.task?.id ?? previous?.taskID,
-            teamID: previous?.teamID,
-            workerID: previous?.workerID,
-            runID: previous?.runID,
-            state: runtime.executionState,
-            updatedAt: Date(),
-            errorMessage: runtime.lastError ?? previous?.errorMessage
-        )
-    }
-
-    private func finishChatRuntime(
-        _ runtime: ChatWorkerRuntime,
-        state: TeamRunState,
-        error: String? = nil
-    ) {
-        runtime.executionState = state
-        runtime.startedAt = nil
-        runtime.lastError = error
-        runtime.dispatchedMode = nil
-        runtime.dispatchedTeamRunID = nil
-        runtime.dispatchedInPlanMode = false
-        updateBackgroundChatState(runtime)
-    }
-
-    /// Where a message goes when it could not be delivered. A drained queue
-    /// entry returns to the head of the queue — writing it into the draft
-    /// would destroy whatever the user typed while waiting.
-    private func stashUnsent(_ text: String, requeue: Bool, preserveDraft: Bool) {
-        if requeue {
-            queuedMessages.insert(text, at: 0)
-            showToast("Kept in queue — reconnect the local agent to send")
-        } else if preserveDraft {
-            draftText = text
-            showToast("Draft kept — reconnect the local agent to send")
-        } else {
-            showToast("Not sent — reconnect the local agent and try again")
-        }
-    }
-
-    /// Sends text to the agent verbatim, without local slash-command matching.
-    /// `execute(_:argument:)` must use this for commands it forwards (like
-    /// /compact) — routing them back through send() would re-match the same
-    /// command and recurse without bound.
-    private func sendRaw(_ text: String) {
-        if isBusy || hasPendingPermission {
-            queuedMessages.append(text)
-            showToast("Queued — sends when this turn finishes")
-            return
-        }
-        guard isAgentOnline,
-              conversationBackend.send(["type": "user_message", "text": text])
-        else {
-            showToast("Reconnect the local agent to run \(text)")
-            return
-        }
-        isBusy = true
-        turnStartedAt = Date()
-        planApprovalPending = false
-        planTodosChangedThisTurn = false
-        clearPendingQuestion()
-        turnDispatchedInPlanMode = false
-        turnDispatchedMode = nil
-        blocks.append(ChatBlock(kind: .user, text: text))
-    }
-
-    func submitDraft() {
-        if isBusy {
-            queueDraft()
-        } else {
-            send(draftText)
-        }
-    }
-
-    /// Append the current direction to the active provider turn. The backend
-    /// stops only the current generation, preserves completed tool results,
-    /// and continues the same turn without an intermediate `turn_done`.
-    func steerDraft() {
-        let text = draftText.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard isBusy,
-              steeringState?.hasPrefix("Stopping") != true,
-              !hasPendingPermission,
-              !text.isEmpty
-        else { return }
-        guard conversationBackend.send(["type": "steer", "text": text]) else {
-            showToast("Reconnect the local agent — the direction was not sent")
-            return
-        }
-        if draftText.trimmingCharacters(in: .whitespacesAndNewlines) == text {
-            draftText = ""
-        }
-        recordPrompt(text)
-        steeringState = "Applying direction…"
-        showToast("Steering the active turn")
-    }
-
-    /// Explicitly retain a message for the next independent turn.
-    func queueDraft() {
-        let text = draftText.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !text.isEmpty else { return }
-        queuedMessages.append(text)
-        if draftText.trimmingCharacters(in: .whitespacesAndNewlines) == text {
-            draftText = ""
-        }
-        recordPrompt(text)
-        showToast("Queued for the next turn")
-    }
-
-    /// Interrupt now, but do not start the replacement turn until the backend
-    /// confirms the old one has fully unwound and persisted its terminal state.
-    func stopAndSendDraft() {
-        let text = draftText.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard isBusy, !hasPendingPermission, !text.isEmpty else { return }
-        guard conversationBackend.send(["type": "interrupt"]) else {
-            showToast("Reconnect the local agent — the active turn could not be stopped")
-            return
-        }
-        computerControl.cancelPendingActions()
-        cancelSimulatorActions(sessionID: currentSessionID)
-        // Scoped: only this conversation is being stopped; a background
-        // worker's in-flight page action keeps its real outcome.
-        browser.cancelPendingActions(ownedBy: currentSessionID)
-        pendingStopAndSend = text
-        if draftText.trimmingCharacters(in: .whitespacesAndNewlines) == text {
-            draftText = ""
-        }
-        recordPrompt(text)
-        steeringState = "Stopping before a new turn…"
-        showToast("Stopping, then sending as a new turn")
-    }
-
-    func removeQueuedMessage(at index: Int) {
-        guard queuedMessages.indices.contains(index) else { return }
-        queuedMessages.remove(at: index)
-        taskWorkers[currentSessionID]?.queuedMessages = queuedMessages
-    }
-
-    private func drainQueuedMessages() {
-        guard !isBusy, !hasPendingPermission, !planApprovalPending,
-              pendingUserQuestion == nil, !queuedMessages.isEmpty else {
-            return
-        }
-        guard isAgentOnline else { return }
-        // A queued message was composed before any attachments added while it
-        // waited; those belong to the user's next explicit send.
-        let message = queuedMessages.removeFirst()
-        taskWorkers[currentSessionID]?.queuedMessages = queuedMessages
-        send(
-            message,
-            preservingDraftOnFailure: false,
-            requeueingOnFailure: true,
-            includeAttachments: false
-        )
-    }
-
-    func previousPrompt() {
-        guard !promptHistory.isEmpty, !isBusy else { return }
-        if promptHistoryCursor == nil {
-            // Stash the unsent draft so leaving history restores it.
-            stashedDraft = draftText
-        }
-        let next = min((promptHistoryCursor ?? -1) + 1, promptHistory.count - 1)
-        promptHistoryCursor = next
-        draftText = promptHistory[next]
-    }
-
-    func nextPrompt() {
-        guard let cursor = promptHistoryCursor, !isBusy else { return }
-        if cursor <= 0 {
-            promptHistoryCursor = nil
-            draftText = stashedDraft ?? ""
-            stashedDraft = nil
-        } else {
-            promptHistoryCursor = cursor - 1
-            draftText = promptHistory[cursor - 1]
-        }
-    }
-
-    /// True while the composer is showing a recalled history entry, so arrow
-    /// keys keep navigating history instead of moving the caret.
-    var isBrowsingPromptHistory: Bool {
-        promptHistoryCursor != nil
-    }
-
-    private func resetHistoryCursorIfEdited() {
-        guard let cursor = promptHistoryCursor,
-              promptHistory.indices.contains(cursor),
-              draftText != promptHistory[cursor]
-        else { return }
-        promptHistoryCursor = nil
-        stashedDraft = nil
-    }
-
-    func copyMessage(_ text: String) {
-        NSPasteboard.general.clearContents()
-        NSPasteboard.general.setString(text, forType: .string)
-        showToast("Message copied")
-    }
-
-    /// Copies a completed assistant response from its authoritative source.
-    /// Rendering never depends on the currently expanded portion of a code
-    /// block or table, so a visually collapsed response still copies in full.
-    func copyResponse(_ source: String, format: ResponseCopyFormat) {
-        let text = ResponseCopyPayload.text(from: source, format: format)
-        guard !text.isEmpty else { return }
-        NSPasteboard.general.clearContents()
-        NSPasteboard.general.setString(text, forType: .string)
-        showToast("Copied as \(format.title)")
-    }
-
-    func useAsDraft(_ text: String) {
-        guard !isBusy, !hasPendingPermission else {
-            showToast("Finish the active action before reusing a message")
-            return
-        }
-        draftText = text
-        showToast("Message moved to the composer")
-    }
-
-    func canRegenerate(_ block: ChatBlock) -> Bool {
-        !isBusy
-            && !hasPendingPermission
-            && block.kind == .assistant
-            && !block.isStreaming
-            && blocks.last(where: { $0.kind == .assistant })?.id == block.id
-    }
-
-    func retryLastResponse() {
-        guard !isBusy, !hasPendingPermission,
-              blocks.contains(where: { $0.kind == .user })
-        else { return }
-        guard conversationBackend.send(["type": "retry_last"]) else {
-            showToast("Reconnect the local agent before retrying")
-            return
-        }
-        pendingRetry = true
-        isBusy = true
-        turnStartedAt = Date()
-        planApprovalPending = false
-        planTodosChangedThisTurn = false
-        planReadyThisTurn = false
-        clearPendingQuestion()
-        turnDispatchedInPlanMode = selectedMode == .plan
-        turnDispatchedMode = selectedMode
-        sessionOverview.emit(.status(
-            status: .running,
-            reason: nil,
-            at: Self.sessionTimestamp
-        ))
-        beginSessionFileCapture()
-        showToast("Regenerating the last response")
-    }
-
-    func stop() {
-        if let pendingTurn = pendingChatTurns[currentSessionID] {
-            let queuedRunID = taskConversationStates[currentSessionID]?.runID
-            pendingTurn.cancel()
-            pendingChatTurns.removeValue(forKey: currentSessionID)
-            pendingChatTurnTokens.removeValue(forKey: currentSessionID)
-            chatAdmissionQueue.remove(currentSessionID)
-            if let runtime = taskWorkers[currentSessionID] {
-                finishChatRuntime(runtime, state: .cancelled)
-            } else {
-                let previous = taskConversationStates[currentSessionID]
-                taskConversationStates[currentSessionID] = TaskConversationState(
-                    sessionID: currentSessionID,
-                    taskID: previous?.taskID,
-                    teamID: previous?.teamID,
-                    workerID: previous?.workerID,
-                    runID: previous?.runID,
-                    state: .cancelled,
-                    updatedAt: Date()
-                )
-            }
-            isBusy = false
-            turnStartedAt = nil
-            turnDispatchedMode = nil
-            turnDispatchedTeamRunID = nil
-            turnDispatchedInPlanMode = false
-            showToast("Removed the queued run")
-            if let queuedRunID {
-                Task { [weak self] in
-                    try? await self?.backend.patch(
-                        "/api/runs/\(queuedRunID)/queue", body: ["action": "cancel"],
-                        as: OrchestrationRun.self
-                    )
-                }
-            }
-            return
-        }
-        // If the interrupt cannot be delivered the run is still live on the
-        // agent; leave the busy state to recoverFromLostConnection(), the one
-        // place that reconciles cards and spinners after a drop.
-        guard conversationBackend.send(["type": "interrupt"]) else {
-            showToast("Reconnect the local agent — the run could not be stopped")
-            return
-        }
-        computerControl.cancelPendingActions()
-        cancelSimulatorActions(sessionID: currentSessionID)
-        browser.cancelPendingActions(ownedBy: currentSessionID)
-        pendingRetry = false
-        steeringState = "Stopping the current run…"
-        showToast("Stopping the current run")
-    }
-
-    var hasRunningWorkForQuit: Bool {
-        !pendingChatTurns.isEmpty || terminal.hasForegroundJob || Self.shouldWarnBeforeQuit(
-            isBusy: isBusy,
-            hasPendingPermission: hasPendingPermission,
-            currentSessionID: currentSessionID,
-            orchestrationState: orchestrationState,
-            taskConversationStates: taskConversationStates,
-            liveWorkerSessionIDs: Set(
-                taskWorkers.values.compactMap { runtime in
-                    runtime.process.isRunning ? runtime.sessionID : nil
-                }
-            )
-        )
-    }
-
-    /// A terminal durable run is authoritative for the current chat. Team
-    /// workers intentionally remain alive between turns, so neither their
-    /// process nor a stale pre-completion snapshot proves work is still active.
-    static func shouldWarnBeforeQuit(
-        isBusy: Bool,
-        hasPendingPermission: Bool,
-        currentSessionID: String,
-        orchestrationState: TeamRunState?,
-        taskConversationStates: [String: TaskConversationState],
-        liveWorkerSessionIDs: Set<String>
-    ) -> Bool {
-        // These foreground flags are set synchronously when a new solo or team
-        // turn begins, before a fresh orchestration event can replace the last
-        // run's terminal state.
-        if isBusy || hasPendingPermission {
-            return true
-        }
-        if let orchestrationState, !orchestrationState.isTerminal {
-            return true
-        }
-        let currentRunIsTerminal = orchestrationState?.isTerminal == true
-        return taskConversationStates.contains { sessionID, snapshot in
-            guard liveWorkerSessionIDs.contains(sessionID), !snapshot.state.isTerminal else {
-                return false
-            }
-            // Completion events can arrive before older foreground flags and
-            // task snapshots are cleared. Do not turn those stale values into
-            // a destructive-looking quit confirmation.
-            return sessionID != currentSessionID || !currentRunIsTerminal
-        }
-    }
-
-    func stopRunningWorkForQuit(completion: @escaping @MainActor () -> Void) {
-        terminal.terminate()
-        for pendingTurn in pendingChatTurns.values { pendingTurn.cancel() }
-        pendingChatTurns.removeAll()
-        pendingChatTurnTokens.removeAll()
-        chatAdmissionQueue = ChatAdmissionQueue()
-        for runtime in taskWorkers.values {
-            _ = runtime.service.send(["type": "interrupt"])
-        }
-        _ = backend.send(["type": "interrupt"])
-        computerControl.cancelPendingActions()
-        cancelSimulatorActions()
-        simulatorControl.detachAll()
-        browser.cancelPendingActions()
-        guard hasRunningWorkForQuit else {
-            completion()
-            return
-        }
-        Task { @MainActor in
-            // Give every worker a bounded window to append its interrupted
-            // task state and terminal event before shutdown stops processes.
-            for _ in 0..<20 {
-                if !hasRunningWorkForQuit { break }
-                try? await Task.sleep(for: .milliseconds(150))
-            }
-            completion()
-        }
-    }
-
-    /// Security-sensitive services must be locked synchronously before either
-    /// the updater or AppKit starts tearing down the rest of the process.
-    func lockSensitiveServicesForShutdown() {
-        walletGateway.lock()
-    }
-
-    func authorizeWalletSession() async {
-        let authorized = await walletGateway.authorizeSession()
-        refreshWalletCapabilities()
-        showToast(authorized ? "Locus Vault unlocked for this session" : "Locus Vault could not be unlocked")
-    }
-
-    func lockWalletSession() {
-        walletGateway.lock()
-        refreshWalletCapabilities()
-    }
-
-    // MARK: - Slash commands
-
-    func execute(_ command: SlashCommand, argument: String) {
-        switch command.action {
-        case .clearChat, .newSession:
-            requestClearChat()
-        case .setMode(let mode):
-            selectedMode = mode
-            showToast(mode == .ask ? "Just Chat is on" : "Switched to \(mode.title) mode")
-        case .selectModel:
-            selectModelMatching(argument)
-        case .browseModels:
-            modelLibraryPresented = true
-        case .refreshModels:
-            Task {
-                await refreshMetadata()
-                showToast("Models refreshed")
-            }
-        case .createCheckpoint:
-            if argument.isEmpty {
-                checkpointPresented = true
-            } else {
-                createCheckpoint(title: argument)
-            }
-        case .manageCheckpoints:
-            checkpointPresented = true
-        case .reviewChanges:
-            selectInspectorTab(.changes)
-        case .openPreview:
-            selectInspectorTab(.preview)
-        case .addContext:
-            addContext()
-        case .exportSession:
-            exportCurrentSession()
-        case .chooseWorkspace:
-            chooseWorkspace()
-        case .newWorkspace:
-            createWorkspace()
-        case .openSettings:
-            settingsPresented = true
-        case .setPermissionMode(let mode):
-            setPermissionMode(mode)
-        case .showShortcuts:
-            shortcutsPresented = true
-        case .showHelp:
-            let help = SlashCommand.all.map(\.helpLine).joined(separator: "\n")
-            blocks.append(ChatBlock(kind: .note, text: "Available commands:\n\(help)"))
-        case .copyLastResponse:
-            if let last = blocks.last(where: { $0.kind == .assistant && !$0.text.isEmpty }) {
-                copyMessage(last.text)
-            } else {
-                showToast("No response to copy yet")
-            }
-        case .retryLastResponse:
-            retryLastResponse()
-        case .stopRun:
-            stop()
-        case .compact:
-            sendRaw("/compact")
-        case .remember:
-            if argument.isEmpty {
-                settingsPage = .knowledge
-                settingsPresented = true
-            } else {
-                rememberConfirmationText = argument
-            }
-        case .setThinkingVisibility:
-            setThinkingVisibility(argument)
-        }
-    }
-
-    private func selectModelMatching(_ argument: String) {
-        guard !argument.isEmpty else {
-            let installed = models.map(\.name).joined(separator: "\n")
-            blocks.append(ChatBlock(
-                kind: .note,
-                text: installed.isEmpty
-                    ? "No Ollama models installed yet — use /models to browse Hugging Face."
-                    : "Installed models:\n\(installed)\n\nUse /model <name> to switch."
-            ))
-            return
-        }
-        let lower = argument.lowercased()
-        let match = models.first { $0.name.caseInsensitiveCompare(argument) == .orderedSame }
-            ?? models.first { $0.name.lowercased().contains(lower) }
-        if let match {
-            selectModel(match.name)
-        } else {
-            showToast("No installed model matches “\(argument)”")
-        }
-    }
-
-    private func setThinkingVisibility(_ argument: String) {
-        let trimmed = argument.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        guard !trimmed.isEmpty else {
-            let options = ThinkingVisibility.allCases
-                .map { "/thinking \($0.rawValue) — \($0.detail)" }
-                .joined(separator: "\n")
-            blocks.append(ChatBlock(
-                kind: .note,
-                text: "Thinking is \(thinkingVisibility.title.lowercased()).\n\(options)"
-            ))
-            return
-        }
-        guard let visibility = ThinkingVisibility(rawValue: trimmed) else {
-            showToast("Use /thinking hidden, collapsed, or expanded")
-            return
-        }
-        thinkingVisibility = visibility
-        showToast("Thinking \(visibility.title.lowercased())")
-    }
-
-    // MARK: - Rewind
-
-    func canRewind(to block: ChatBlock) -> Bool {
-        block.kind == .user && !isBusy && !hasPendingPermission
-    }
-
-    /// Restores the conversation to the state just before a user message and
-    /// places that message back in the composer for editing — Claude Code's
-    /// per-message rewind, built on the checkpoint mechanism.
-    func rewind(to block: ChatBlock) {
-        guard canRewind(to: block),
-              let index = blocks.firstIndex(where: { $0.id == block.id })
-        else { return }
-        let firstLine = block.text
-            .components(separatedBy: .newlines)
-            .first.map { String($0.prefix(42)) } ?? "message"
-        let checkpoint = SessionCheckpoint(
-            id: UUID(),
-            title: "Rewind point — \(firstLine)",
-            createdAt: Date(),
-            blocks: Array(blocks.prefix(upTo: index)),
-            todos: [],
-            contextFiles: contextFiles,
-            workspacePath: workspacePath,
-            model: selectedModel,
-            activePlan: nil
-        )
-        pendingRewindDraft = block.text
-        restore(checkpoint)
-    }
-
-    func decide(requestID: String, decision: String) {
-        // UI tests drive the prompt against a dead socket; applying the
-        // decision locally is what lets the panel advance and dismiss.
-        if !isUITesting {
-            guard conversationBackend.send([
-                "type": "permission_decision",
-                "request_id": requestID,
-                "decision": decision,
-            ]) else {
-                showToast("The agent disconnected before the decision was sent")
-                return
-            }
-        }
-        if let index = blocks.lastIndex(where: { $0.tool?.requestID == requestID }) {
-            blocks[index].tool?.status = decision == "deny" ? .denied : .running
-        }
-        if let runtime = taskWorkers[currentSessionID] {
-            runtime.executionState = .running
-            updateBackgroundChatState(runtime)
-        }
-        if orchestrationRunID != nil {
-            orchestrationState = .running
-            updateTaskConversation(state: .running, event: [:])
-        }
-    }
-
-    func selectModel(_ model: String) {
-        rememberManualModelRoute(accountID: activeAccount?.id, model: model)
-        if isBusy {
-            pendingProviderSwitch = (activeAccount?.id, model)
-            showToast("Switching to \(model) after this turn")
-            return
-        }
-        guard backend.send(["type": "set_model", "model": model]) else {
-            showToast("Reconnect before switching models")
-            return
-        }
-        showToast("Switching to \(model)")
-    }
-
-    /// Routes the session to a model, switching providers when the model comes
-    /// from a different source than the one in use.
-    ///
-    /// `account` nil means local Ollama. Switching providers replaces the
-    /// agent's client, which it refuses to do mid-turn — so a switch requested
-    /// during a run is held and applied when the turn finishes.
-    func selectModel(account: ProviderAccount?, model: String) {
-        rememberManualModelRoute(accountID: account?.id, model: model)
-        let sameSource = account?.id.uuidString == settings.activeAccountID
-        guard !sameSource else {
-            if let account {
-                rememberPreferredModel(model, for: account)
-                // The window belongs to the model, not to the account. Sending
-                // only `set_model` left the agent budgeting against the model we
-                // just switched away from: a Claude account moved from a
-                // 1,000,000-token model to a 200,000-token one kept metering
-                // against 1,000,000 and would not compact until five times over
-                // the real window, failing every request past it.
-                Task { [weak self] in
-                    await self?.applyProvider(announce: false)
-                    // After the provider call, so the transcript records the
-                    // switch against the model the agent has actually adopted.
-                    self?.selectModel(model)
-                }
-            } else {
-                selectModel(model)
-            }
-            return
-        }
-        if let account, !account.hasKey {
-            showToast("Add an API key for \(account.displayName) in Settings")
-            settingsPresented = true
-            return
-        }
-        guard !isBusy else {
-            pendingProviderSwitch = (account?.id, model)
-            showToast("Switching to \(account?.displayName ?? "local Ollama") after this turn")
-            return
-        }
-        applyProviderSwitch(accountID: account?.id, model: model)
-    }
-
-    /// A provider switch that arrived mid-turn, applied once the agent is idle.
-    private var pendingProviderSwitch: (accountID: UUID?, model: String)?
-
-    func applyPendingProviderSwitchIfNeeded() {
-        guard let pending = pendingProviderSwitch else { return }
-        pendingProviderSwitch = nil
-        applyProviderSwitch(accountID: pending.accountID, model: pending.model)
-    }
-
-    private func applyProviderSwitch(accountID: UUID?, model: String) {
-        // The route is committed before the agent has accepted it, because the
-        // request body is built from these fields.
-        let previousAccountID = settings.activeAccountID
-        let previousProvider = settings.provider
-        if let accountID, let account = providerAccounts.first(where: { $0.id == accountID }) {
-            rememberPreferredModel(model, for: account)
-            settings.activeAccountID = accountID.uuidString
-            settings.provider = .remote
-        } else {
-            settings.activeAccountID = nil
-            settings.provider = .ollama
-        }
-        persistSettings()
-        Task {
-            guard await applyProvider() else {
-                // The agent kept the provider it had. Leaving the new account
-                // committed would leave the app pointing at an account that
-                // never connected while every turn still runs on the old
-                // route — visible as an account paired with another
-                // provider's model.
-                settings.activeAccountID = previousAccountID
-                settings.provider = previousProvider
-                persistSettings()
-                return
-            }
-            // The remote provider adopts its configured model as it connects;
-            // the local runtime keeps whatever it had, so name it explicitly.
-            if accountID == nil, !model.isEmpty, model != selectedModel {
-                selectModel(model)
-            }
-            persistCurrentWorkspaceProfile()
-        }
-    }
-
-    private func rememberPreferredModel(_ model: String, for account: ProviderAccount) {
-        guard modelBelongsToAccount(model, account: account) else { return }
-        guard let index = providerAccounts.firstIndex(where: { $0.id == account.id }),
-              providerAccounts[index].preferredModel != model
-        else { return }
-        providerAccounts[index].preferredModel = model
-        persistProviderAccounts()
-    }
-
-    private func modelBelongsToAccount(_ model: String, account: ProviderAccount) -> Bool {
-        let trimmed = model.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return false }
-        if account.kind != .custom {
-            return ProviderModelFilter.matches(kind: account.kind, name: trimmed)
-        }
-        if case .connected = accountStatus[account.id],
-           let reported = accountModels[account.id], !reported.isEmpty
-        {
-            return reported.contains { $0.caseInsensitiveCompare(trimmed) == .orderedSame }
-        }
-        let routed = agentProfiles.filter { $0.route.accountID == account.id }.map(\.model)
-        return routed.isEmpty || routed.contains {
-            $0.caseInsensitiveCompare(trimmed) == .orderedSame
-        }
-    }
-
-    // MARK: - Agents and teams
-
-    func testAgentProfileConnection(_ profile: AgentProfile) async -> String {
-        switch profile.route {
-        case .localOllama:
-            guard let url = URL(string: lastOllamaHost + "/api/tags") else {
-                return "The local Ollama URL is invalid."
-            }
-            do {
-                let (data, response) = try await ProxyRuntime.shared.urlSession.data(from: url)
-                guard (200..<300).contains((response as? HTTPURLResponse)?.statusCode ?? -1) else {
-                    return "Ollama did not accept the connection."
-                }
-                let object = try? JSONSerialization.jsonObject(with: data) as? [String: Any]
-                let names = (object?["models"] as? [[String: Any]] ?? []).compactMap {
-                    $0["name"] as? String
-                }
-                if !profile.model.isEmpty,
-                   !names.contains(where: { $0.caseInsensitiveCompare(profile.model) == .orderedSame })
-                {
-                    return "Connected, but that exact model is not installed in Ollama."
-                }
-                return "Connected to local Ollama."
-            } catch {
-                return "Could not connect to Ollama: \(error.localizedDescription)"
-            }
-        case .providerAccount(let id):
-            guard let account = providerAccounts.first(where: { $0.id == id }) else {
-                return "That provider account is unavailable."
-            }
-            let result = await ProviderModelCatalog.fetch(for: account)
-            accountModels[id] = result.models
-            accountStatus[id] = result.status
-            guard result.status.isHealthy else { return result.status.summary }
-            if account.kind.listsModels,
-               !profile.model.isEmpty,
-               !result.models.contains(where: { $0.caseInsensitiveCompare(profile.model) == .orderedSame })
-            {
-                return "Connected, but the exact model was not in this account's catalog."
-            }
-            return result.status.summary
-        }
-    }
-
-    /// Builds an in-memory manifest for one run. Provider credentials are
-    /// included only in the WebSocket payload and are never written into the
-    /// profile/team stores or transcript.
-    func teamManifest(for text: String, teamID: UUID? = nil) -> [String: Any]? {
-        let mention = TeamMentionResolver.selection(
-            in: text,
-            profiles: agentProfiles,
-            teams: agentTeams
-        )
-        let team = teamID.flatMap { requested in
-            agentTeams.first(where: { $0.id == requested })
-        } ?? mention.team
-            ?? selectedAgentTeam
-            ?? mention.agent.flatMap { agent in
-                agentTeams.first(where: { $0.memberIDs.contains(agent.id) })
-            }
-        guard let team else { return nil }
-        let errors = AgentTeamValidation.errors(team: team, profiles: agentProfiles)
-        guard errors.isEmpty else {
-            showToast(errors[0])
-            return nil
-        }
-        let routeErrors = AgentTeamValidation.routeErrors(
-            team: team,
-            profiles: agentProfiles,
-            accounts: providerAccounts,
-            accountModels: accountModels
-        )
-        guard routeErrors.isEmpty else {
-            showToast(routeErrors[0])
-            return nil
-        }
-        let members = team.memberIDs.compactMap { id in agentProfiles.first(where: { $0.id == id }) }
-        for profile in members {
-            guard let accountID = profile.route.accountID else { continue }
-            guard teamRoutingConsentAccountIDs.contains(accountID) else {
-                let label = providerAccounts.first(where: { $0.id == accountID })?.displayName
-                    ?? "hosted account"
-                showToast("Allow automatic team routing for \(label) in Agents & Teams")
-                return nil
-            }
-        }
-        let routes: [[String: Any]] = members.compactMap { profile in
-            var route: [String: Any]
-            switch profile.route {
-            case .localOllama:
-                route = [
-                    "provider": "ollama",
-                    "host": lastOllamaHost,
-                ]
-            case .providerAccount(let accountID):
-                guard let account = providerAccounts.first(where: { $0.id == accountID }),
-                      account.hasKey
-                else { return nil }
-                if account.kind == .chatGPT {
-                    route = [
-                        "provider": "chatgpt",
-                        "account_id": account.id.uuidString,
-                        "codex_home_id": account.codexHomeIdentifier,
-                        "account_label": account.displayName,
-                        // Always sent: a missing field means "keep the
-                        // current server-side value", not "use the default".
-                        "native_mode": account.codexNativeModeEnabled,
-                        "web_search": account.codexWebSearchEnabled,
-                        "reasoning_effort": account.codexReasoningEffortValue,
-                    ]
-                } else {
-                    route = [
-                        "provider": "remote",
-                        "base_url": account.resolvedBaseURL,
-                        "api_key": CredentialStore.get(account: account.credentialAccount) ?? "",
-                        "auth_style": account.kind.authStyle,
-                        "account_kind": account.kind.rawValue,
-                        "lists_models": account.kind.listsModels,
-                        "account_label": account.displayName,
-                    ]
-                }
-            }
-            var entry: [String: Any] = [
-                "id": profile.id.uuidString,
-                "name": profile.name,
-                "model": profile.model,
-                "role": profile.role.rawValue,
-                "instructions": profile.instructions,
-                "capabilities": profile.capabilityTags,
-                "access_ceiling": profile.accessCeiling.rawValue,
-                "timeout_seconds": profile.timeoutSeconds,
-                "token_limit": profile.tokenLimit,
-                "metering": route["provider"] as? String == "chatgpt"
-                    ? AgentMetering.selfHosted.rawValue
-                    : profile.metering.rawValue,
-                "route": route,
-            ]
-            if let behavior = encodedJSONObject(profile.resolvedBehavior) {
-                entry["behavior"] = behavior
-            }
-            if let rate = profile.inputCostPerMillion { entry["input_cost_per_million"] = rate }
-            if let rate = profile.outputCostPerMillion { entry["output_cost_per_million"] = rate }
-            if let policy = profile.mcpPolicy,
-               let data = try? JSONEncoder().encode(policy),
-               let value = try? JSONSerialization.jsonObject(with: data)
-            {
-                entry["mcp_policy"] = value
-            }
-            return entry
-        }
-        guard routes.count == members.count else {
-            showToast("A team member's provider account is unavailable")
-            return nil
-        }
-        var teamPayload: [String: Any] = [
-            "id": team.id.uuidString,
-            "name": team.name,
-            "member_ids": team.memberIDs.map(\.uuidString),
-            "use_managed_worktree": team.useManagedWorktree,
-            "parallel_writers": team.resolvedParallelWriters,
-            // One approval releases the complete plan. Individual jobs and
-            // models do not introduce additional dispatch confirmations.
-            "dispatch_approval_mode": DispatchApprovalMode.preview.rawValue,
-            "routing_mode": team.resolvedRoutingMode.rawValue,
-            "routing_weights": [
-                "quality": team.resolvedRoutingWeights.quality,
-                "reliability": team.resolvedRoutingWeights.reliability,
-                "privacy": team.resolvedRoutingWeights.privacy,
-                "latency": team.resolvedRoutingWeights.latency,
-                "cost": team.resolvedRoutingWeights.cost,
-            ],
-            "evaluation_tags": team.evaluationTags ?? [],
-            "maximum_estimated_cost": team.maximumEstimatedCost ?? 0,
-            "swarm_policy": [
-                "version": team.resolvedSwarmPolicy.version,
-                "engine": team.resolvedSwarmPolicy.engine.rawValue,
-                "delegation_mode": team.resolvedSwarmPolicy.delegationMode.rawValue,
-                "sizing_mode": team.resolvedSwarmPolicy.sizingMode.rawValue,
-                "max_total_agents": team.resolvedSwarmPolicy.maxTotalAgents,
-                "max_depth": team.resolvedSwarmPolicy.maxDepth,
-            ],
-            "budget": [
-                "max_jobs": team.budget.maxJobs,
-                "max_rounds": team.budget.maxRounds,
-                "max_model_calls": team.budget.maxModelCalls,
-                "max_concurrent_calls": team.budget.maxConcurrentCalls,
-                "max_metered_tokens": team.budget.maxMeteredTokens,
-                "call_budget_mode": team.budget.callBudgetMode.rawValue,
-            ],
-        ]
-        if let id = team.dispatcherID { teamPayload["dispatcher_id"] = id.uuidString }
-        if let id = team.fallbackDispatcherID { teamPayload["fallback_dispatcher_id"] = id.uuidString }
-        if let id = team.defaultWriterID { teamPayload["default_writer_id"] = id.uuidString }
-        var manifest: [String: Any] = [
-            "run_id": UUID().uuidString,
-            "team": teamPayload,
-            "profiles": routes,
-        ]
-        if let id = mention.agent?.id { manifest["forced_agent_id"] = id.uuidString }
-        return manifest
-    }
-
-    func exportOrchestration(_ runID: String, includeContent: Bool) async {
-        do {
-            let value: [String: JSONValue] = try await backend.get(
-                "/api/orchestrations/\(runID)/export",
-                query: [URLQueryItem(
-                    name: "include_content", value: includeContent ? "true" : "false"
-                )],
-                as: [String: JSONValue].self
-            )
-            let encoder = JSONEncoder()
-            encoder.outputFormatting = [.prettyPrinted, .sortedKeys, .withoutEscapingSlashes]
-            let data = try encoder.encode(value)
-            let panel = NSSavePanel()
-            panel.nameFieldStringValue = "\(runID).locusrun"
-            panel.allowedContentTypes = [UTType(filenameExtension: "locusrun") ?? .json]
-            guard panel.runModal() == .OK, let url = panel.url else { return }
-            try data.write(to: url, options: .atomic)
-            showToast(includeContent ? "Run exported with visible content" : "Redacted run exported")
-        } catch {
-            showToast("Could not export run: \(error.localizedDescription)")
-        }
-    }
-
-    func exportRunToOTLP(_ runID: String, includeContent: Bool = false) async {
-        guard settings.otlpExportEnabled,
-              !settings.otlpEndpoint.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
-        else { return }
-        do {
-            let _: SimpleActionResponse = try await backend.post(
-                "/api/runs/\(runID)/otlp",
-                body: [
-                    "endpoint": settings.otlpEndpoint,
-                    "authorization": settings.otlpAuthorization,
-                    "include_content": includeContent,
-                ],
-                as: SimpleActionResponse.self
-            )
-            await refreshOrchestrationRuns(select: runID)
-        } catch {
-            await refreshOrchestrationRuns(select: runID)
-            showToast("Telemetry export failed: \(error.localizedDescription)")
-        }
-    }
-
-    private func exportOrchestrationToOTLP(_ runID: String) async {
-        let sample = AppSettings.clampOTLPSamplingRate(settings.otlpSamplingRate)
-        guard sample >= 1 || (sample > 0 && Double.random(in: 0..<1) < sample) else { return }
-        await exportRunToOTLP(runID, includeContent: false)
-    }
-
-    func pauseOrchestration(_ runID: String) {
-        orchestrationAction(path: "/api/orchestrations/\(runID)/pause", runID: runID)
-    }
-
-    func cancelOrchestration(_ runID: String) {
-        let transport = orchestrationBackend(for: runID)
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let _: SimpleActionResponse = try await transport.post(
-                    "/api/orchestrations/\(runID)/cancel",
-                    body: [:],
-                    timeout: 30,
-                    as: SimpleActionResponse.self
-                )
-                if orchestrationRunID == runID {
-                    pendingDispatchPlan = nil
-                    orchestrationState = .cancelled
-                    steeringState = "Stopping the team run…"
-                    updateTaskConversation(
-                        state: .cancelled,
-                        event: ["run_id": runID]
-                    )
-                }
-                showToast("Stopping the team run")
-                await refreshOrchestrationRuns(select: runID)
-            } catch {
-                showToast("Could not stop the team run: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func discardOrchestration(_ runID: String) {
-        orchestrationAction(path: "/api/orchestrations/\(runID)/discard", runID: nil)
-    }
-
-    func cleanupOrchestrationCheckout(_ run: OrchestrationRun) {
-        guard let taskID = run.taskID else { return }
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let response: TaskMutationResponse = try await backend.delete(
-                    "/api/tasks/\(taskID)", as: TaskMutationResponse.self
-                )
-                if activeTaskRecord?.id == taskID {
-                    activeTaskRecord = response.task
-                    taskHasChanges = false
-                }
-                showToast("Managed checkout archived with a restorable snapshot")
-                await refreshOrchestrationRuns(select: run.id)
-            } catch {
-                showToast(error.localizedDescription)
-            }
-        }
-    }
-
-    func setOrchestrationPinned(_ run: OrchestrationRun, pinned: Bool) {
-        Task {
-            do {
-                let updated: OrchestrationRun = try await backend.patch(
-                    "/api/orchestrations/\(run.id)",
-                    body: ["pinned": pinned],
-                    as: OrchestrationRun.self
-                )
-                selectedOrchestrationRun = updated
-                await refreshOrchestrationRuns(select: updated.id)
-            } catch {
-                showToast("Could not update run: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func resumeOrchestration(_ run: OrchestrationRun) {
-        guard teamRunPresentation(for: run.id, durable: run).canRecover else {
-            showToast("That team run is not paused or interrupted")
-            return
-        }
-        guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-              let manifest = teamManifest(for: run.request, teamID: teamID)
-        else {
-            showToast("Repair the team, models, or hosted consent before resuming")
-            return
-        }
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let assessment: RunRecoveryAssessment = try await backend.post(
-                    "/api/orchestrations/\(run.id)/recovery-assessment",
-                    body: ["manifest": manifest],
-                    as: RunRecoveryAssessment.self
-                )
-                guard assessment.canResume else {
-                    showToast(assessment.repairChecklist.first ?? "This run cannot be resumed")
-                    return
-                }
-                orchestrationAction(
-                    path: "/api/orchestrations/\(run.id)/resume",
-                    body: ["manifest": manifest],
-                    runID: run.id
-                )
-            } catch {
-                showToast("Could not assess recovery: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func retryOrchestrationJob(_ attempt: AgentJobAttempt, in run: OrchestrationRun) {
-        guard teamRunPresentation(for: run.id, durable: run).canRecover else {
-            showToast("Pause the team run before retrying a job")
-            return
-        }
-        guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-              let manifest = teamManifest(for: run.request, teamID: teamID)
-        else {
-            showToast("Repair the team before retrying this job")
-            return
-        }
-        orchestrationAction(
-            path: "/api/orchestrations/\(run.id)/jobs/\(attempt.jobID)/retry",
-            body: ["manifest": manifest],
-            runID: run.id
-        )
-    }
-
-    func stopOrchestrationBranch(_ attempt: AgentJobAttempt, in run: OrchestrationRun) {
-        guard teamRunPresentation(for: run.id, durable: run).isActivelyOwned,
-              !isCodingAttempt(attempt, in: run)
-        else {
-            showToast("Only an active read-only branch can be stopped")
-            return
-        }
-        guard let node = encodedAgentNode(attempt.resolvedNodeID) else {
-            showToast("That agent branch has an invalid identity")
-            return
-        }
-        orchestrationAction(
-            path: "/api/orchestrations/\(run.id)/agents/\(node)/stop",
-            runID: run.id
-        )
-    }
-
-    func retryOrchestrationBranch(_ attempt: AgentJobAttempt, in run: OrchestrationRun) {
-        guard teamRunPresentation(for: run.id, durable: run).canRecover else {
-            showToast("Pause the team run before retrying a branch")
-            return
-        }
-        guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-              let manifest = teamManifest(for: run.request, teamID: teamID),
-              let node = encodedAgentNode(attempt.resolvedNodeID)
-        else {
-            showToast("Repair the team before retrying this branch")
-            return
-        }
-        orchestrationAction(
-            path: "/api/orchestrations/\(run.id)/agents/\(node)/retry",
-            body: ["manifest": manifest],
-            runID: run.id
-        )
-    }
-
-    func runOrchestrationWithLocus(_ run: OrchestrationRun) {
-        guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-              let manifest = teamManifest(for: run.request, teamID: teamID)
-        else {
-            showToast("Repair the team before continuing with Locus")
-            return
-        }
-        orchestrationAction(
-            path: "/api/orchestrations/\(run.id)/run-with-locus",
-            body: ["manifest": manifest],
-            runID: run.id
-        )
-    }
-
-    private func encodedAgentNode(_ nodeID: String) -> String? {
-        var allowed = CharacterSet.alphanumerics
-        allowed.insert(charactersIn: "-._~")
-        return nodeID.addingPercentEncoding(withAllowedCharacters: allowed)
-    }
-
-    func reassignOrchestrationJob(
-        _ attempt: AgentJobAttempt,
-        in run: OrchestrationRun,
-        to profile: AgentProfile
-    ) {
-        guard teamRunPresentation(for: run.id, durable: run).canRecover else {
-            showToast("Pause the team run before reassigning a job")
-            return
-        }
-        guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-              let manifest = teamManifest(for: run.request, teamID: teamID)
-        else {
-            showToast("Repair the team before reassigning this job")
-            return
-        }
-        orchestrationAction(
-            path: "/api/orchestrations/\(run.id)/jobs/\(attempt.jobID)/reassign",
-            body: ["manifest": manifest, "agent_id": profile.id.uuidString],
-            runID: run.id
-        )
-    }
-
-    func reassignmentCandidates(
-        for attempt: AgentJobAttempt,
-        in run: OrchestrationRun
-    ) -> [AgentProfile] {
-        guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-              let team = agentTeams.first(where: { $0.id == teamID })
-        else { return [] }
-        return team.memberIDs.compactMap { id in
-            agentProfiles.first(where: { $0.id == id })
-        }.filter { profile in
-            !profile.accessCeiling.canWrite
-                && profile.id.uuidString != attempt.agentID
-                && (attempt.role != "reviewer" || profile.role == .reviewer)
-        }.sorted { $0.name.localizedCaseInsensitiveCompare($1.name) == .orderedAscending }
-    }
-
-    func isCodingAttempt(_ attempt: AgentJobAttempt, in run: OrchestrationRun) -> Bool {
-        guard case .object(let plan) = run.checkpoint?.state["plan"],
-              case .array(let jobs) = plan["jobs"]
-        else { return false }
-        return jobs.contains { value in
-            guard case .object(let job) = value else { return false }
-            return job["id"]?.string == attempt.jobID && job["kind"]?.string == "writer"
-        }
-    }
-
-    func replayOrchestration(_ run: OrchestrationRun) {
-        startOrchestrationCopy(run, action: "replay")
-    }
-
-    func duplicateOrchestration(_ run: OrchestrationRun) {
-        startOrchestrationCopy(run, action: "duplicate")
-    }
-
-    private func startOrchestrationCopy(_ run: OrchestrationRun, action: String) {
-        guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-              let manifest = teamManifest(for: run.request, teamID: teamID)
-        else {
-            showToast("Repair the team, models, or hosted consent before continuing")
-            return
-        }
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let response: OrchestrationMutationResponse = try await backend.post(
-                    "/api/orchestrations/\(run.id)/\(action)",
-                    body: ["manifest": manifest],
-                    timeout: 30,
-                    as: OrchestrationMutationResponse.self
-                )
-                orchestrationRunID = response.runID
-                await refreshOrchestrationRuns(select: response.runID)
-            } catch {
-                showToast(error.localizedDescription)
-            }
-        }
-    }
-
-    func decideDispatch(_ action: String, editedPlan: DispatchPlan? = nil) {
-        guard let runID = orchestrationRunID else { return }
-        if action == "cancel" {
-            cancelOrchestration(runID)
-            return
-        }
-        var payload: [String: Any] = ["type": "dispatch_decision", "run_id": runID, "action": action]
-        if let editedPlan, let value = encodedJSONObject(editedPlan) { payload["plan"] = value }
-        guard orchestrationBackend(for: runID).send(payload) else {
-            showToast("The dispatch decision could not be delivered")
-            return
-        }
-        pendingDispatchPlan = nil
-        if let runtime = taskWorkers[currentSessionID] {
-            runtime.executionState = action == "redispatch" ? .dispatching : .running
-            updateBackgroundChatState(runtime)
-        }
-        if action == "redispatch" {
-            orchestrationState = .dispatching
-            dispatcherValidationReason = nil
-            if var activity = dispatcherActivity {
-                activity.state = .running
-                activity.output = "Creating a new dispatcher plan…"
-                activity.startedAt = Date()
-                dispatcherActivity = activity
-            }
-            updateTaskConversation(state: .dispatching, event: ["run_id": runID])
-        }
-    }
-
-    func dispatchPlanErrors(_ plan: DispatchPlan) -> [String] {
-        let runTeamID = selectedOrchestrationRun?.teamID.flatMap(UUID.init(uuidString:))
-        guard let team = runTeamID.flatMap({ id in agentTeams.first(where: { $0.id == id }) })
-            ?? selectedAgentTeam
-        else { return ["The selected team is unavailable."] }
-        let profiles = Dictionary(uniqueKeysWithValues: agentProfiles.map { ($0.id, $0) })
-        let budget = plan.budget ?? team.budget
-        var errors: [String] = []
-        if plan.jobs.isEmpty || plan.jobs.count > budget.maxJobs {
-            errors.append("The plan must contain 1…\(budget.maxJobs) jobs.")
-        }
-        let ids = plan.jobs.map(\.id)
-        if ids.contains(where: { $0.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty })
-            || Set(ids).count != ids.count
-        {
-            errors.append("Every job needs a unique ID.")
-        }
-        let known = Set(ids)
-        for job in plan.jobs {
-            guard let agentID = UUID(uuidString: job.agentID),
-                  team.memberIDs.contains(agentID), let profile = profiles[agentID]
-            else {
-                errors.append("Job \(job.id) uses an unavailable team member.")
-                continue
-            }
-            if job.kind == "writer" && !profile.accessCeiling.canWrite {
-                errors.append("Coding job \(job.id) requires a write-capable team member.")
-            }
-            if job.kind != "writer" && profile.accessCeiling.canWrite {
-                errors.append("Write-capable team members may only own coding jobs.")
-            }
-            if job.kind == "reviewer" && profile.role != .reviewer {
-                errors.append("Reviewer jobs require a Reviewer profile.")
-            }
-            if let role = job.requiredRole, !role.isEmpty, profile.role.rawValue != role {
-                errors.append("Job \(job.id) requires the \(role) role.")
-            }
-            if !Set(job.capabilityTags ?? []).isSubset(of: Set(profile.capabilityTags)) {
-                errors.append("Job \(job.id) requires capabilities its agent does not have.")
-            }
-            if job.dependencies.contains(job.id)
-                || job.dependencies.contains(where: { !known.contains($0) })
-            {
-                errors.append("Job \(job.id) has an invalid dependency.")
-            }
-        }
-        let writerJobs = plan.jobs.filter { $0.kind == "writer" }
-        if writerJobs.isEmpty {
-            errors.append("The plan must contain at least one coding job.")
-        }
-        var visiting: Set<String> = []
-        var visited: Set<String> = []
-        let dependencies = Dictionary(uniqueKeysWithValues: plan.jobs.map { ($0.id, $0.dependencies) })
-        func visit(_ id: String) -> Bool {
-            if visited.contains(id) { return false }
-            if visiting.contains(id) { return true }
-            visiting.insert(id)
-            for dependency in dependencies[id] ?? [] where visit(dependency) { return true }
-            visiting.remove(id)
-            visited.insert(id)
-            return false
-        }
-        if ids.contains(where: visit) { errors.append("The job graph contains a dependency cycle.") }
-        let kindByID = Dictionary(uniqueKeysWithValues: plan.jobs.map { ($0.id, $0.kind) })
-        for job in plan.jobs {
-            if job.kind == "specialist",
-               job.dependencies.contains(where: { kindByID[$0] != "specialist" })
-            {
-                errors.append("Specialists may depend only on specialist jobs.")
-            }
-            if job.kind == "writer",
-               job.dependencies.contains(where: {
-                   guard let kind = kindByID[$0] else { return false }
-                   return kind != "specialist" && kind != "writer"
-               })
-            {
-                errors.append("Coding jobs may depend only on specialists or earlier coding jobs.")
-            }
-        }
-        func transitivelyDepends(_ jobID: String, on targetID: String, seen: inout Set<String>) -> Bool {
-            guard seen.insert(jobID).inserted else { return false }
-            for dependency in dependencies[jobID] ?? [] {
-                if dependency == targetID { return true }
-                if transitivelyDepends(dependency, on: targetID, seen: &seen) { return true }
-            }
-            return false
-        }
-        if writerJobs.count > 1 {
-            for leftIndex in 0..<(writerJobs.count - 1) {
-                for rightIndex in (leftIndex + 1)..<writerJobs.count {
-                    var leftSeen: Set<String> = []
-                    var rightSeen: Set<String> = []
-                    let ordered = transitivelyDepends(
-                        writerJobs[leftIndex].id,
-                        on: writerJobs[rightIndex].id,
-                        seen: &leftSeen
-                    ) || transitivelyDepends(
-                        writerJobs[rightIndex].id,
-                        on: writerJobs[leftIndex].id,
-                        seen: &rightSeen
-                    )
-                    if !ordered {
-                        errors.append("Every pair of coding jobs must be ordered by a dependency.")
-                    }
-                }
-            }
-        }
-        let minimumModelCalls = plan.jobs.count + 2 + (budget.maxRounds > 1 ? 1 : 0)
-        if budget.maxModelCalls < minimumModelCalls {
-            errors.append("This plan needs at least \(minimumModelCalls) model calls for its jobs, synthesis, and possible lead revision.")
-        }
-        if budget.maxConcurrentCalls > budget.maxModelCalls {
-            errors.append("Concurrent calls cannot exceed the model-call budget.")
-        }
-        for id in team.memberIDs {
-            if let accountID = profiles[id]?.route.accountID,
-               !teamRoutingConsentAccountIDs.contains(accountID)
-            {
-                errors.append("Hosted automatic-routing consent is missing.")
-                break
-            }
-        }
-        return Array(Set(errors)).sorted()
-    }
-
-    private func orchestrationAction(
-        path: String,
-        body: [String: Any] = [:],
-        runID: String?
-    ) {
-        let transport = runID.map(orchestrationBackend(for:)) ?? backend
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let _: SimpleActionResponse = try await transport.post(
-                    path, body: body, timeout: 30, as: SimpleActionResponse.self
-                )
-                await refreshOrchestrationRuns(select: runID)
-            } catch {
-                showToast(error.localizedDescription)
-            }
-        }
-    }
-
-    func openWorkspaceMemorySource(_ memory: WorkspaceMemory) {
-        if let runID = memory.sourceRunID, !runID.isEmpty {
-            selectInspectorTab(.agents)
-            inspectorCollapsed = false
-            Task { await loadOrchestrationRun(runID) }
-            return
-        }
-        if let sessionID = memory.sourceSessionID,
-           let session = sessions.first(where: { $0.id == sessionID })
-        {
-            resume(session)
-        } else {
-            showToast("The source chat is no longer available")
-        }
-    }
-
-    var currentLandingCheckCommands: [String] {
-        workspaceProfiles.first(where: {
-            SessionSummary.canonicalWorkspacePath($0.path)
-                == SessionSummary.canonicalWorkspacePath(workspacePath)
-        })?.resolvedLandingCheckCommands ?? []
-    }
-
-    func saveLandingCheckCommands(_ commands: [String]) {
-        let clean = Array(commands.map {
-            String($0.trimmingCharacters(in: .whitespacesAndNewlines).prefix(500))
-        }.filter { !$0.isEmpty }.prefix(8))
-        touchWorkspaceProfile(workspacePath)
-        if let index = workspaceProfiles.firstIndex(where: {
-            SessionSummary.canonicalWorkspacePath($0.path)
-                == SessionSummary.canonicalWorkspacePath(workspacePath)
-        }) {
-            workspaceProfiles[index].landingCheckCommands = clean
-            persistWorkspaceProfiles()
-        }
-    }
-
-    func presentScheduleEditor(task: ScheduledTask? = nil, prompt: String? = nil) {
-        if let task {
-            scheduleEditorDraft = ScheduleEditorDraft(task: task)
-            return
-        }
-        var draft = ScheduleEditorDraft()
-        draft.prompt = prompt ?? draftText
-        draft.workspaceRoot = sessionInfo?.workspaceRoot ?? workspacePath
-        draft.mode = selectedMode
-        draft.executionEnvironment = sessionInfo?.environment?["type"] == "worktree"
-            ? .worktree : .local
-        if let team = selectedAgentTeam {
-            draft.runner = .team
-            draft.teamID = team.id.uuidString
-            draft.teamName = team.name
-        } else {
-            draft.runner = .solo
-        }
-        if let account = activeAccount {
-            draft.provider = account.kind == .chatGPT ? "chatgpt" : "remote"
-            draft.providerAccountID = account.id.uuidString
-            draft.model = routedModel(for: account)
-        } else {
-            draft.provider = "ollama"
-            draft.model = selectedModel
-        }
-        scheduleEditorDraft = draft
-    }
-
-    func rememberScheduleWorkspace(_ url: URL) -> String? {
-        guard workspaceAccess.rememberAndActivate(url) else { return nil }
-        return url.standardizedFileURL.resolvingSymlinksInPath().path
-    }
-
-    func openSchedules() {
-        activityCenterSection = .schedules
-        activityCenterPresented = true
-        Task { @MainActor [weak self] in
-            await self?.refreshScheduledTasks()
-        }
-    }
-
-    private func scheduleConfigurationIssue(for draft: ScheduleEditorDraft) -> String? {
-        let name = draft.name.trimmingCharacters(in: .whitespacesAndNewlines)
-        let prompt = draft.prompt.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !name.isEmpty else { return "Add a schedule name" }
-        guard !prompt.isEmpty else { return "Add a prompt" }
-        guard FileManager.default.fileExists(atPath: draft.workspaceRoot),
-              workspaceAccess.activateStored(path: draft.workspaceRoot)
-        else { return "Choose an available workspace folder" }
-        guard !draft.model.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty,
-              draft.model != "No model"
-        else { return "Choose an available model" }
-        if draft.ruleKind == .once, draft.oneTimeDate <= Date() {
-            return "Choose a future date and time"
-        }
-        if draft.ruleKind == .interval {
-            let seconds = draft.intervalEvery * [
-                .minutes: 60, .hours: 3_600, .days: 86_400, .weeks: 604_800,
-            ][draft.intervalUnit, default: 0]
-            guard seconds >= 900 else { return "Custom intervals must be at least 15 minutes" }
-        }
-        if draft.runner == .team {
-            guard let id = draft.teamID.flatMap(UUID.init(uuidString:)),
-                  agentTeams.contains(where: { $0.id == id })
-            else { return "Choose an available team" }
-        }
-        if draft.provider != "ollama" {
-            guard let id = draft.providerAccountID.flatMap(UUID.init(uuidString:)),
-                  let account = providerAccounts.first(where: { $0.id == id })
-            else { return "Choose an available model account" }
-            let expected = account.kind == .chatGPT ? "chatgpt" : "remote"
-            guard draft.provider == expected else { return "The selected model account changed" }
-        }
-        return nil
-    }
-
-    private func scheduleConfigurationIssue(for task: ScheduledTask) -> String? {
-        guard FileManager.default.fileExists(atPath: task.workspaceRoot),
-              workspaceAccess.activateStored(path: task.workspaceRoot)
-        else { return "The workspace bookmark is no longer available" }
-        guard !task.model.isEmpty else { return "The configured model is no longer available" }
-        if task.runner == .team {
-            guard let id = task.teamID.flatMap(UUID.init(uuidString:)),
-                  let team = agentTeams.first(where: { $0.id == id })
-            else { return "The configured team no longer exists" }
-            if let issue = AgentTeamValidation.errors(team: team, profiles: agentProfiles).first {
-                return issue
-            }
-        }
-        if task.provider == "ollama" {
-            if !installedLocalModels.isEmpty,
-               !installedLocalModels.contains(where: { $0.name == task.model }) {
-                return "The configured local model is no longer installed"
-            }
-        } else {
-            guard let id = task.providerAccountID.flatMap(UUID.init(uuidString:)),
-                  let account = providerAccounts.first(where: { $0.id == id })
-            else { return "The configured model account no longer exists" }
-            let expected = account.kind == .chatGPT ? "chatgpt" : "remote"
-            guard task.provider == expected else { return "The configured model account changed" }
-            if let catalog = accountModels[id], !catalog.isEmpty, !catalog.contains(task.model) {
-                return "The configured model is no longer offered by this account"
-            }
-        }
-        return nil
-    }
-
-    func updateQueuedRun(_ run: OrchestrationRun, action: String) {
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let _: OrchestrationRun = try await backend.patch(
-                    "/api/runs/\(run.id)/queue", body: ["action": action],
-                    as: OrchestrationRun.self
-                )
-                if let sessionID = run.sessionID {
-                    if action == "cancel" {
-                        pendingChatTurns[sessionID]?.cancel()
-                        pendingChatTurns.removeValue(forKey: sessionID)
-                        pendingChatTurnTokens.removeValue(forKey: sessionID)
-                        chatAdmissionQueue.remove(sessionID)
-                        if let runtime = taskWorkers[sessionID] {
-                            finishChatRuntime(runtime, state: .cancelled)
-                        }
-                    } else {
-                        chatAdmissionQueue.move(sessionID, action: action)
-                    }
-                }
-                await refreshActivityRuns()
-            } catch { showToast(error.localizedDescription) }
-        }
-    }
-
-    func retryRun(_ run: OrchestrationRun) {
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let retry: OrchestrationRun = try await backend.post(
-                    "/api/runs/\(run.id)/retry", body: [:], as: OrchestrationRun.self
-                )
-                guard let sessionID = retry.sessionID,
-                      let session = sessions.first(where: { $0.id == sessionID }),
-                      let workspace = retry.workspaceRoot ?? session.workspacePath
-                else {
-                    showToast("The original chat or workspace is unavailable")
-                    return
-                }
-                guard let worker = await ensureChatWorker(
-                    for: sessionID,
-                    workspaceRoot: workspace,
-                    provider: retry.manifest?["provider"]?.string,
-                    providerAccountID: retry.manifest?["provider_account_id"]?.string,
-                    model: retry.manifest?["model"]?.string
-                ) else {
-                    showToast("The original chat worker could not be started")
-                    return
-                }
-                let retryMode = retry.manifest?["mode"]?.string
-                    .flatMap { WorkMode.canonical($0) } ?? .work
-                worker.reservedRunID = retry.id
-                worker.dispatchedMode = retryMode
-                worker.executionState = .queued
-                taskConversationStates[sessionID] = TaskConversationState(
-                    sessionID: sessionID,
-                    taskID: retry.taskID,
-                    teamID: retry.teamID,
-                    workerID: retry.workerID,
-                    runID: retry.id,
-                    state: .queued,
-                    updatedAt: Date()
-                )
-                guard await waitForChatExecutionSlot(worker) else { return }
-                let _: OrchestrationRun = try await backend.patch(
-                    "/api/runs/\(retry.id)/queue", body: ["action": "admit"],
-                    as: OrchestrationRun.self
-                )
-                var request: [String: Any] = [
-                    "type": "user_message",
-                    "text": Self.decoratedPrompt(
-                        retry.request,
-                        mode: retryMode,
-                        chatAttachments: [],
-                        contextFiles: [],
-                        restoredTranscriptContext: nil
-                    ),
-                    "mode": retryMode.rawValue,
-                    "run_id": retry.id,
-                ]
-                if let config = encodedJSONObject(primaryAgentBehavior) {
-                    request["agent_config"] = config
-                }
-                if retry.runKind == "team",
-                   let teamID = retry.teamID.flatMap(UUID.init(uuidString:)),
-                   var manifest = teamManifest(for: retry.request, teamID: teamID) {
-                    manifest["run_id"] = retry.id
-                    request["team"] = manifest
-                    worker.dispatchedTeamRunID = retry.id
-                    worker.executionState = .dispatching
-                } else {
-                    worker.executionState = .running
-                    if retry.isSoloSwarm {
-                        request["solo_swarm"] = ["enabled": true]
-                    }
-                }
-                guard worker.service.send(request) else {
-                    finishChatRuntime(worker, state: .failed, error: "The retry could not be delivered")
-                    return
-                }
-                worker.startedAt = Date()
-                updateBackgroundChatState(worker)
-                showToast("Retry queued in \(session.displayTitle)")
-                await refreshActivityRuns()
-            } catch { showToast(error.localizedDescription) }
-        }
-    }
-
-    private func restorePersistedQueuedRuns() {
-        let queued = activityRuns.filter { $0.state == "queued" }.sorted {
-            ($0.queuePosition ?? .max) < ($1.queuePosition ?? .max)
-        }
-        for run in queued where restoredQueuedRunIDs.insert(run.id).inserted {
-            Task { @MainActor [weak self] in
-                await self?.dispatchPersistedQueuedRun(run)
-            }
-        }
-    }
-
-    func dispatchPersistedQueuedRun(_ run: OrchestrationRun) async {  // internal(for: AppModel extension files)
-        guard let sessionID = run.sessionID,
-              let workspace = run.workspaceRoot,
-              let worker = await ensureChatWorker(
-                for: sessionID,
-                workspaceRoot: workspace,
-                provider: run.manifest?["provider"]?.string,
-                providerAccountID: run.manifest?["provider_account_id"]?.string,
-                model: run.manifest?["model"]?.string
-              )
-        else {
-            restoredQueuedRunIDs.remove(run.id)
-            showToast("A saved queued run needs its original chat and workspace")
-            return
-        }
-        let mode = run.manifest?["mode"]?.string.flatMap { WorkMode.canonical($0) } ?? .work
-        worker.reservedRunID = run.id
-        worker.dispatchedMode = mode
-        worker.executionState = .queued
-        taskConversationStates[sessionID] = TaskConversationState(
-            sessionID: sessionID,
-            taskID: run.taskID,
-            teamID: run.teamID,
-            workerID: run.workerID,
-            runID: run.id,
-            state: .queued,
-            updatedAt: Date()
-        )
-        guard await waitForChatExecutionSlot(worker) else { return }
-        do {
-            let _: OrchestrationRun = try await backend.patch(
-                "/api/runs/\(run.id)/queue", body: ["action": "admit"],
-                as: OrchestrationRun.self
-            )
-            var request: [String: Any] = [
-                "type": "user_message",
-                "text": Self.decoratedPrompt(
-                    run.request, mode: mode, chatAttachments: [], contextFiles: [],
-                    restoredTranscriptContext: nil
-                ),
-                "mode": mode.rawValue,
-                "run_id": run.id,
-            ]
-            if let config = encodedJSONObject(primaryAgentBehavior) {
-                request["agent_config"] = config
-            }
-            if run.runKind == "team" {
-                guard let teamID = run.teamID.flatMap(UUID.init(uuidString:)),
-                      var manifest = teamManifest(for: run.request, teamID: teamID) else {
-                    finishChatRuntime(
-                        worker, state: .interrupted,
-                        error: "The saved team configuration needs attention before resuming"
-                    )
-                    return
-                }
-                manifest["run_id"] = run.id
-                request["team"] = manifest
-                worker.dispatchedTeamRunID = run.id
-                worker.executionState = .dispatching
-            } else {
-                worker.executionState = .running
-                if run.isSoloSwarm {
-                    request["solo_swarm"] = ["enabled": true]
-                }
-            }
-            guard worker.service.send(request) else {
-                finishChatRuntime(worker, state: .interrupted, error: "The saved run could not be delivered")
-                return
-            }
-            worker.startedAt = Date()
-            updateBackgroundChatState(worker)
-        } catch {
-            finishChatRuntime(worker, state: .interrupted, error: error.localizedDescription)
-        }
-    }
-
-    func openActivityRun(_ run: OrchestrationRun) {
-        guard let sessionID = run.sessionID,
-              let session = sessions.first(where: { $0.id == sessionID })
-        else { showToast("That chat is no longer available"); return }
-        markActivitySeen(run)
-        activityCenterPresented = false
-        resume(session)
-        Task { await loadOrchestrationRun(run.id) }
-    }
-
-    func openNotification(sessionID: String, runID: String) {
-        activityCenterPresented = false
-        if let session = sessions.first(where: { $0.id == sessionID }) {
-            resume(session)
-        }
-        if !runID.isEmpty {
-            Task { await loadOrchestrationRun(runID) }
-        }
-    }
-
-    func stopActivityRun(_ run: OrchestrationRun) {
-        if run.state == "queued" {
-            updateQueuedRun(run, action: "cancel")
-            return
-        }
-        if let sessionID = run.sessionID, let runtime = taskWorkers[sessionID] {
-            guard runtime.service.send(["type": "interrupt"]) else {
-                showToast("That chat worker could not be reached")
-                return
-            }
-            runtime.executionState = .cancelled
-            updateBackgroundChatState(runtime)
-            showToast("Stopping the selected run")
-            return
-        }
-        cancelOrchestration(run.id)
-    }
-
-    func answerActivityPermission(_ run: OrchestrationRun, decision: String) {
-        guard let sessionID = run.sessionID,
-              let runtime = taskWorkers[sessionID],
-              let event = runtime.pendingForegroundEvent,
-              event["type"] as? String == "permission_request",
-              let requestID = event["request_id"] as? String,
-              runtime.service.send([
-                "type": "permission_decision",
-                "request_id": requestID,
-                "decision": decision,
-              ])
-        else {
-            showToast("Open the chat to review this permission request")
-            return
-        }
-        runtime.pendingForegroundEvent = nil
-        runtime.executionState = .running
-        updateBackgroundChatState(runtime)
-        showToast(decision == "deny" ? "Permission denied" : "Permission granted")
-        Task { await refreshActivityRuns() }
-    }
-
-    func publishLandedWorktree() {
-        guard let task = activeTaskRecord, let branch = task.branch,
-              GitRemoteFeatures.isAvailable else { return }
-        let client = GitClient(workspaceRoot: task.executionPath)
-        isLandingOperationRunning = true
-        Task { @MainActor [weak self] in
-            defer { self?.isLandingOperationRunning = false }
-            do {
-                let upstream = try? await client.run([
-                    "rev-parse", "--abbrev-ref", "--symbolic-full-name", "@{upstream}",
-                ])
-                try await client.run(
-                    GitPushPlan.arguments(
-                        branch: branch,
-                        upstream: upstream?.stdout.trimmingCharacters(in: .whitespacesAndNewlines)
-                    ),
-                    timeout: 120
-                )
-                self?.showToast("Published \(branch)")
-            } catch {
-                self?.showToast("Publish failed; the branch and commit are safe: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func openLandedPullRequest() {
-        guard let task = activeTaskRecord, let branch = task.branch else { return }
-        let client = GitClient(workspaceRoot: task.executionPath)
-        Task { @MainActor [weak self] in
-            guard let remote = try? await client.run(["remote", "get-url", "origin"]),
-                  let url = GitRemoteURL.githubCompareURL(
-                    remote: remote.stdout.trimmingCharacters(in: .whitespacesAndNewlines),
-                    branch: branch
-                  ) else {
-                self?.showToast("The origin remote is not a GitHub repository")
-                return
-            }
-            NSWorkspace.shared.open(url)
-        }
-    }
-
-    var currentExecutionEnvironment: ChatExecutionEnvironment {
-        if let raw = sessionInfo?.environment?["type"],
-           let environment = ChatExecutionEnvironment(rawValue: raw) {
-            return environment
-        }
-        return activeTaskRecord == nil ? .local : .worktree
-    }
-
-    func handoffCurrentChat(to environment: ChatExecutionEnvironment) {
-        guard !isBusy, !hasPendingPermission, !currentSessionID.isEmpty else {
-            showToast("Wait for the current turn before handing off")
-            return
-        }
-        guard environment != currentExecutionEnvironment else { return }
-        let sessionID = currentSessionID
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let response: SessionHandoffResponse = try await conversationBackend.post(
-                    "/api/sessions/\(sessionID)/handoff",
-                    body: ["environment": environment.rawValue, "base_ref": "HEAD"],
-                    timeout: 120,
-                    as: SessionHandoffResponse.self
-                )
-                sessionInfo = response.sessionInfo
-                activeTaskRecord = response.task
-                if let runtime = taskWorkers[sessionID] {
-                    runtime.sessionInfo = response.sessionInfo
-                }
-                gitWorkspace.refreshStatus()
-                await refreshMetadata()
-                showToast(
-                    environment == .worktree
-                        ? "Chat moved to its worktree"
-                        : "Chat and changes moved to Local"
-                )
-            } catch {
-                showToast("Handoff left both checkouts unchanged: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func createBranchForActiveTask(_ rawName: String) {
-        guard let task = activeTaskRecord, !isBusy, !hasPendingPermission else { return }
-        let name = rawName.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !name.isEmpty else { return }
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let response: TaskMutationResponse = try await conversationBackend.post(
-                    "/api/tasks/\(task.id)/branch",
-                    body: ["branch": name],
-                    as: TaskMutationResponse.self
-                )
-                activeTaskRecord = response.task
-                sessionInfo = sessionInfo?.replacingTask(response.task)
-                gitWorkspace.refreshBranch()
-                showToast("Created branch \(name) in the worktree")
-            } catch {
-                showToast("Could not create branch: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func restoreActiveTaskCheckout() {
-        guard let task = activeTaskRecord, !isBusy else { return }
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let response: TaskMutationResponse = try await conversationBackend.post(
-                    "/api/tasks/\(task.id)/restore",
-                    body: [:],
-                    timeout: 120,
-                    as: TaskMutationResponse.self
-                )
-                activeTaskRecord = response.task
-                showToast("Worktree restored")
-            } catch {
-                showToast("Could not restore the worktree: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func restoreWorktree(for session: SessionSummary) {
-        guard let task = session.task else { return }
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let _: TaskMutationResponse = try await backend.post(
-                    "/api/tasks/\(task.id)/restore",
-                    body: [:],
-                    timeout: 120,
-                    as: TaskMutationResponse.self
-                )
-                await refreshMetadata()
-                showToast("Worktree restored")
-            } catch {
-                showToast("Could not restore worktree: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func copyActiveTaskPatch() {
-        guard let task = activeTaskRecord else { return }
-        Task { @MainActor [weak self] in
-            guard let self else { return }
-            do {
-                let response: TaskDetailResponse = try await backend.get(
-                    "/api/tasks/\(task.id)",
-                    as: TaskDetailResponse.self
-                )
-                let pasteboard = NSPasteboard.general
-                pasteboard.clearContents()
-                pasteboard.setString(response.patch, forType: .string)
-                showToast("Copied task patch")
-            } catch {
-                showToast("Could not copy the task patch: \(error.localizedDescription)")
-            }
-        }
-    }
-
-    func openActiveTaskCheckout() {
-        guard let path = activeTaskRecord?.executionPath else { return }
-        NSWorkspace.shared.open(URL(fileURLWithPath: path, isDirectory: true))
-    }
-
-    func revealActiveTaskCheckout() {
-        guard let path = activeTaskRecord?.executionPath else { return }
-        NSWorkspace.shared.activateFileViewerSelecting([URL(fileURLWithPath: path, isDirectory: true)])
-    }
-
-    /// Adds or updates an account. The key is written here rather than in the
-    /// editor so an abandoned sheet leaves nothing behind; `apiKey` nil means
-    /// "keep the saved one".
-    @discardableResult
-    func saveProviderAccount(_ account: ProviderAccount, apiKey: String?) -> Bool {
-        if account.kind != .chatGPT {
-            let effectiveKey = apiKey ?? CredentialStore.get(account: account.credentialAccount) ?? ""
-            if let error = RemoteEndpointTester.securityError(
-                baseURL: account.resolvedBaseURL,
-                apiKey: effectiveKey
-            ) {
-                showToast(error)
-                return false
-            }
-        }
-        var updated = account
-        updated.name = ProviderAccountStore.uniqueName(
-            account.name,
-            kind: account.kind,
-            existing: providerAccounts,
-            excluding: account.id
-        )
-        // Write the credential before publishing the account. Otherwise a disk
-        // or permission failure produces a convincing "Saved" account whose
-        // key never survived, and closing the editor loses the only copy the
-        // user may have of a one-time key.
-        if let apiKey, updated.kind.requiresAPIKey,
-           !providerCredentialWriter(apiKey, updated.credentialAccount)
-        {
-            showToast("Could not save the API key to \(CredentialStore.displayPath)")
-            return false
-        }
-        if let index = providerAccounts.firstIndex(where: { $0.id == updated.id }) {
-            providerAccounts[index] = updated
-        } else {
-            providerAccounts.append(updated)
-        }
-        persistProviderAccounts()
-        forgetAccountCatalog(updated.id)
-        Task {
-            await refreshAccountCatalogs(force: true)
-            // The live agent is holding the old endpoint or key until it is
-            // told otherwise.
-            if updated.id.uuidString == settings.activeAccountID {
-                await applyProvider(announce: false)
-            }
-        }
-        showToast("Saved \(updated.displayName)")
-        return true
-    }
-
-    /// Removes an account, its key, and — if it was the one in use — the
-    /// routing that depended on it.
-    func removeProviderAccount(_ account: ProviderAccount) {
-        providerAccounts.removeAll { $0.id == account.id }
-        CredentialStore.remove(account: account.credentialAccount)
-        persistProviderAccounts()
-        forgetAccountCatalog(account.id)
-        guard account.id.uuidString == settings.activeAccountID else {
-            showToast("Removed \(account.displayName)")
-            return
-        }
-        if isBusy {
-            pendingProviderSwitch = (nil, "")
-            showToast("Removed \(account.displayName) — local Ollama takes over after this turn")
-        } else {
-            applyProviderSwitch(accountID: nil, model: "")
-            showToast("Removed \(account.displayName) — using local Ollama")
-        }
-    }
-
-    /// Deletes the stored key. When it belongs to the account in use the
-    /// agent is told at once: it holds the key in memory, so leaving it be
-    /// would keep spending a credential the user just revoked.
-    func removeProviderAccountKey(_ account: ProviderAccount) {
-        CredentialStore.remove(account: account.credentialAccount)
-        accountStatus[account.id] = .noKey
-        forgetAccountCatalog(account.id)
-        guard account.id.uuidString == settings.activeAccountID else { return }
-        // Sends an empty key, which the agent treats as "clear it". Held until
-        // the turn finishes when one is running, because /api/provider refuses
-        // mid-turn — and a dropped revocation is the one failure here that
-        // costs the user money.
-        if isBusy {
-            pendingProviderSwitch = (account.id, account.preferredModel)
-            showToast("Key removed — the agent drops it when this turn finishes")
-        } else {
-            Task { await applyProvider(announce: false) }
-        }
-    }
-
-    /// Records that the endpoint rejected this account's key, so Settings and
-    /// the picker can say so instead of leaving the user to guess.
-    func noteAccountKeyRejected() {
-        guard let account = activeAccount else { return }
-        accountStatus[account.id] = .keyRejected
-    }
-
-    func activateInstalledModel(_ reference: String) async {
-        await refreshMetadata()
-        let lowerReference = reference.lowercased()
-        // Ollama lists HF pulls as "hf.co/owner/repo:QUANT". Prefer the exact
-        // name, then the owner-qualified repository, and only fall back to the
-        // bare repo name when it matches a single installed model.
-        let repoID = lowerReference
-            .replacingOccurrences(of: "hf.co/", with: "")
-            .split(separator: ":")
-            .first.map(String.init) ?? lowerReference
-        let repoName = repoID.split(separator: "/").last.map(String.init) ?? repoID
-        var match = models.first { $0.name.caseInsensitiveCompare(reference) == .orderedSame }
-            ?? models.first { $0.name.lowercased().contains(repoID) }
-        if match == nil {
-            let candidates = models.filter { $0.name.lowercased().contains(repoName) }
-            match = candidates.count == 1 ? candidates.first : nil
-        }
-        guard let match else {
-            showToast("Model installed — refresh the model list to select it")
-            return
-        }
-        selectModel(match.name)
-        switch HuggingFaceVariant.fit(
-            bytes: match.size,
-            physicalMemory: ProcessInfo.processInfo.physicalMemory
-        ) {
-        case .fits:
-            showToast("Installed and selected \(match.name)")
-        case .tight:
-            showToast("Installed \(match.name) — it will use most of this Mac's memory")
-        case .exceeds:
-            showToast("Installed \(match.name) — likely too large for this Mac")
-        }
-    }
-
-    /// Hides an Ollama model from Locus without touching its downloaded files.
-    /// The complete Ollama list stays in memory so Settings can restore it.
-    func removeLocalModelFromLocus(_ model: ModelInfo) {
-        guard !isLocalModelHidden(model.name) else { return }
-        settings.hiddenLocalModels.append(model.name)
-        settings.hiddenLocalModels.sort { $0.localizedCaseInsensitiveCompare($1) == .orderedAscending }
-        localModels = visibleLocalModels(in: installedLocalModels)
-        if activeAccount == nil { models = localModels }
-        showToast("Removed \(model.name) from Locus — it is still installed")
-    }
-
-    func restoreLocalModelToLocus(_ model: ModelInfo) {
-        settings.hiddenLocalModels.removeAll {
-            $0.caseInsensitiveCompare(model.name) == .orderedSame
-        }
-        localModels = visibleLocalModels(in: installedLocalModels)
-        if activeAccount == nil { models = localModels }
-        showToast("Restored \(model.name) to Locus")
-    }
-
-    /// Permanently asks Ollama to remove the model's downloaded data. The UI
-    /// owns the confirmation because this operation cannot be undone by Locus.
-    func deleteLocalModelFromComputer(_ model: ModelInfo) async {
-        do {
-            try await LocalModelManagement.delete(ollamaHost: ollamaHost, model: model.name)
-        } catch {
-            showToast("Could not delete \(model.name): \(error.localizedDescription)")
-            return
-        }
-
-        installedLocalModels.removeAll {
-            $0.name.caseInsensitiveCompare(model.name) == .orderedSame
-        }
-        settings.hiddenLocalModels.removeAll {
-            $0.caseInsensitiveCompare(model.name) == .orderedSame
-        }
-        localModels = visibleLocalModels(in: installedLocalModels)
-        if activeAccount == nil {
-            models = localModels
-            if selectedModel.caseInsensitiveCompare(model.name) == .orderedSame,
-               let replacement = localModels.first
-            {
-                selectModel(replacement.name)
-            }
-        }
-        showToast("Deleted \(model.name) from this Mac")
-    }
-
     func requestClearChat() {
         guard (!isBusy && !hasPendingPermission) || taskWorkers[currentSessionID] != nil else {
             showToast("Finish or stop the active run before clearing")
@@ -6881,7 +3971,7 @@ final class AppModel: ObservableObject {
         drainQueuedMessages()
     }
 
-    private func clearPendingQuestion() {
+    func clearPendingQuestion() {
         pendingUserQuestion = nil
         capturedQuestionThisTurn = nil
     }
@@ -8090,7 +5180,7 @@ final class AppModel: ObservableObject {
         return task
     }
 
-    private func cancelSimulatorActions(sessionID: String? = nil) {
+    func cancelSimulatorActions(sessionID: String? = nil) {
         let requestIDs = pendingSimulatorActions.compactMap { requestID, pending in
             sessionID == nil || pending.sessionID == sessionID ? requestID : nil
         }
@@ -8393,7 +5483,7 @@ final class AppModel: ObservableObject {
         ])
     }
 
-    private func refreshWalletCapabilities() {
+    func refreshWalletCapabilities() {
         sendWalletCapability(to: backend)
         for runtime in taskWorkers.values {
             sendWalletCapability(to: runtime.service)
@@ -8528,7 +5618,7 @@ final class AppModel: ObservableObject {
         return (try? await backend.get("/api/health", as: HealthResponse.self)) != nil
     }
 
-    private func ensureChatWorker(
+    func ensureChatWorker(
         for requestedSessionID: String,
         workspaceRoot: String,
         provider: String? = nil,
@@ -9099,7 +6189,7 @@ final class AppModel: ObservableObject {
         )
     }
 
-    private static func decoratedPrompt(
+    static func decoratedPrompt(
         _ text: String,
         mode: WorkMode,
         chatAttachments: [ChatAttachment],
@@ -9208,7 +6298,7 @@ final class AppModel: ObservableObject {
         return sections.joined(separator: "\n\n")
     }
 
-    private func updateTaskConversation(
+    func updateTaskConversation(
         state: TeamRunState,
         event: [String: Any],
         taskID: String? = nil
@@ -9271,7 +6361,7 @@ final class AppModel: ObservableObject {
         )
     }
 
-    private func activateSessionOverview(_ info: SessionInfo, reset: Bool = false) {
+    func activateSessionOverview(_ info: SessionInfo, reset: Bool = false) {
         let model = sessionOverviewModel(for: info)
         let initial = SessionState.empty(
             workspacePath: info.workspaceRoot ?? info.cwd,
@@ -9516,7 +6606,7 @@ final class AppModel: ObservableObject {
     }
 
     /// Opens the window in which file activity is attributed to this run.
-    private func beginSessionFileCapture() {
+    func beginSessionFileCapture() {
         synchronizeSessionIdentity()
         fileCaptureSessionID = sessionOverview.activeSessionID
         fileCaptureStartedAt = Self.sessionTimestamp
@@ -10902,7 +7992,7 @@ final class AppModel: ObservableObject {
         }
     }
 
-    private func touchWorkspaceProfile(_ path: String) {
+    func touchWorkspaceProfile(_ path: String) {
         let path = SessionSummary.canonicalWorkspacePath(path)
         if let index = workspaceProfiles.firstIndex(where: {
             SessionSummary.canonicalWorkspacePath($0.path) == path
@@ -10948,7 +8038,7 @@ final class AppModel: ObservableObject {
         }
     }
 
-    private func persistCurrentWorkspaceProfile() {
+    func persistCurrentWorkspaceProfile() {
         // Before the agent reports a session, workspacePath falls back to the
         // home directory — which must never be recorded as a real workspace.
         guard sessionInfo != nil else { return }
@@ -11021,14 +8111,14 @@ final class AppModel: ObservableObject {
         return (installed, nil)
     }
 
-    private func persistWorkspaceProfiles() {
+    func persistWorkspaceProfiles() {
         guard persistenceEnabled else { return }
         if let data = try? JSONEncoder().encode(workspaceProfiles) {
             UserDefaults.standard.set(data, forKey: "Locus.workspaceProfiles")
         }
     }
 
-    private func recordPrompt(_ text: String) {
+    func recordPrompt(_ text: String) {
         promptHistory.removeAll { $0 == text }
         promptHistory.insert(text, at: 0)
         promptHistory = Array(promptHistory.prefix(50))
@@ -11069,7 +8159,7 @@ final class AppModel: ObservableObject {
         )
     }
 
-    private func persistSettings() {
+    func persistSettings() {
         guard persistenceEnabled else { return }
         if let data = try? JSONEncoder().encode(settings) {
             UserDefaults.standard.set(data, forKey: "Locus.settings")
